% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{AllEqual}
\alias{AllEqual}
\title{Check if all values in a vector are the same}
\usage{
AllEqual(x = NULL, y = NULL)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Check if all values in a vector are the same
}
\examples{
loonR::AllEqual(c(1,1,1,1,1))
loonR::AllEqual(c(1,2,1,1,1))
loonR::AllEqual(c(1,2,1)c(1,2,1))
}
