% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{biomarker_discovery_by_CV}
\alias{biomarker_discovery_by_CV}
\title{Title}
\usage{
biomarker_discovery_by_CV(
  normal.samples,
  tumor.samples,
  expression.df,
  n = 100,
  train.proportion = 0.7,
  train.auc.cutoff = 0.65,
  fc = 1,
  validate.auc.cutoff = 0.7,
  candidate.fre = 50,
  candidate.mean.auc = 0.7,
  seed = 500,
  experiment.name = "Tumor",
  ca19 = -1,
  final.train.samples = -1
)
}
\arguments{
\item{n}{CV times. Default 100 times.}

\item{train.auc.cutoff}{Cutoff}

\item{fc}{Fold change cutoff in train dataset}

\item{validate.auc.cutoff}{Cutoff}

\item{candidate.fre}{Frequence}

\item{candidate.mean.auc}{Cutoff}

\item{seed}{Default 500}

\item{ca19}{CA19-9}

\item{final.train.samples}{Vector. Final split train sample. If specified, will not randomly select train samples.}
}
\value{

}
\description{
Title
}
