% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{build.best.random.forest}
\alias{build.best.random.forest}
\title{Try different mtry and select the best fitted model}
\usage{
build.best.random.forest(rf.df, group, ntree = 500, seed = 111, scale = TRUE)
}
\arguments{
\item{rf.df}{Row is sample}

\item{ntree}{Default 500}

\item{seed}{Default 111}

\item{scale}{If perform scale}
}
\description{
Try different mtry and select the best fitted model
}
