% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{build.coxregression.model}
\alias{build.coxregression.model}
\title{Title}
\usage{
build.coxregression.model(d.frame, status, time, seed = 666, scale = TRUE)
}
\arguments{
\item{d.frame}{Data.frame --- Row: sample, Column: gene expression}

\item{time}{OS, DFS, RFS et al.....}

\item{seed}{Default 666}
}
\value{

}
\description{
Title
}
