% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{build.lassoOrRidge.regression}
\alias{build.lassoOrRidge.regression}
\title{Build lasso or ridge regression}
\usage{
build.lassoOrRidge.regression(
  df,
  group,
  seed = 666,
  scale = TRUE,
  alpha = 1,
  lambda = NA,
  nfolds = 10,
  type.measure = "auc"
)
}
\arguments{
\item{seed}{Default 666}

\item{scale}{Default TRUE}

\item{alpha}{1 for lasso, 0 for ridge}

\item{lambda}{}

\item{nfolds}{Default 10, number of folds}

\item{type.measure}{Default auc. Refer https://www.rdocumentation.org/packages/glmnet/versions/4.1-3/topics/cv.glmnet}
}
\value{

}
\description{
Build lasso or ridge regression
}
\examples{
data("LIRI")
reg.res <- loonR::build.lassoOrRidge.regression(LIRI[,3:5],LIRI$status)

}
