% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{build.randomforest.model}
\alias{build.randomforest.model}
\title{Build a random forests model}
\usage{
build.randomforest.model(df, group, seed = 666, scale = TRUE)
}
\arguments{
\item{df}{Row is sample}

\item{scale}{Default TRUE}
}
\description{
Build a random forests model
}
