% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{build.SVM}
\alias{build.SVM}
\title{Build Support Vector Machines}
\usage{
build.SVM(df, group, seed = 666, scale = TRUE)
}
\arguments{
\item{df}{row is sample}

\item{scale}{Default TRUE}
}
\description{
Build Support Vector Machines
}
