% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MachineLearningPlatform.H2O.R
\name{buildEnsembleModel}
\alias{buildEnsembleModel}
\title{Build ensemble model based on h2o platform}
\usage{
buildEnsembleModel(
  train.df = NULL,
  train.group = NULL,
  val.df = NULL,
  val.group = NULL,
  candidates = NULL,
  nfolds = 5,
  seed = 1,
  type = "run",
  metalearner_algorithm = "glm",
  glm = TRUE,
  xg = TRUE,
  dl = TRUE,
  nb = TRUE,
  gbm = TRUE,
  rf = TRUE
)
}
\arguments{
\item{train.df}{row is sample, column is feature}

\item{train.group}{a vector}

\item{val.df}{Default NULL, not to evaluate in validation dataset. row is sample, column is feature}

\item{val.group}{a vector}

\item{candidates}{if not specify candidates feature, all column names will be used}

\item{nfolds}{Default 5. Used in internal model construction}

\item{seed}{Default 1}

\item{type}{Description for this run. Default "run"}

\item{metalearner_algorithm}{Default glm. Could be "AUTO", "deeplearning", "drf", "gbm", "glm", "naivebayes", "xgboost". If AUTO,  (GLM with non negative weights; if validation_frame is present, a lambda search is performed}

\item{glm}{TRUE. Default included glm}

\item{xg}{TRUE. Default included xgboost}

\item{dl}{TRUE. Default included deep learning}

\item{nb}{TRUE. Default included NaiveBayes}

\item{gbm}{TRUE. Default included gbm}

\item{rf}{TRUE. Default included random forests}
}
\value{
A list include fits, training risk score, aucs
}
\description{
Build ensemble model based on h2o platform
}
