% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{buildGRangesObj}
\alias{buildGRangesObj}
\title{Build Genomic Ranges from data.frame}
\usage{
buildGRangesObj(
  df,
  keep.extra.columns = T,
  seqnames.field = "chr",
  start.field = "start",
  end.field = "end",
  starts.in.df.are.0based = FALSE,
  strand.field = "strand"
)
}
\arguments{
\item{keep.extra.columns}{Default TRUE}

\item{seqnames.field}{Default "chr"}

\item{start.field}{Default "start"}

\item{end.field}{Default "end"}

\item{starts.in.df.are.0based}{Default FALSE}

\item{strand.field}{Default "strand"}
}
\value{

}
\description{
Build Genomic Ranges from data.frame
}
