% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{cal_cnvScore}
\alias{cal_cnvScore}
\title{Cal_CNV score}
\usage{
cal_cnvScore(file_path = NULL, save_path = NULL)
}
\arguments{
\item{file_path}{infercnv.observations.txt}

\item{save_path}{save path for cnv score}
}
\description{
Cal_CNV score
}
\examples{
cal_cnvScore("infercnv.observations.txt",  "cnv.score.rdata")
}
