% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNF.R
\name{calculateNMI}
\alias{calculateNMI}
\title{Ranks each features by NMI (normalized mutual information) based on their clustering assingments}
\usage{
calculateNMI(data.list, wall)
}
\arguments{
\item{data.list}{A list, pls specify the name. Element is data.frame (Row is sample, column is feature.). dataL <- list( mRNA=t(mRNA.snf.df), miRNA=t(miRNA.snf.df), Met=t(methylation.snf.df), CNV=t(cnv.snf.df) ).}

\item{wall}{}
}
\description{
Ranks each features by NMI (normalized mutual information) based on their clustering assingments
}
