% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug.R
\name{celline_drug_group_auc_analysis}
\alias{celline_drug_group_auc_analysis}
\title{Analyze if a durg sensitivity is correlated with group/subtype}
\usage{
celline_drug_group_auc_analysis(
  drug.sen.df = NULL,
  group = NULL,
  group.name = "Group",
  compare.method = "wilcox.test"
)
}
\arguments{
\item{group.name}{Default Group}

\item{compare.method}{Default wilcox.test}
}
\value{

}
\description{
Analyze if a durg sensitivity is correlated with group/subtype
}
