% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR.R
\name{checkHousekeepingCombination}
\alias{checkHousekeepingCombination}
\title{Iterate all the combination of housekeeping genes}
\usage{
checkHousekeepingCombination(
  raw.ct.df,
  housekeeping.ct.df,
  group,
  alternative = "less",
  p.cutoff = 0.05,
  difference.cutoff = 0,
  sep = ", "
)
}
\arguments{
\item{raw.ct.df}{row is sample}

\item{housekeeping.ct.df}{row is sample}

\item{group}{Second shuld be TRUE}

\item{alternative}{Default "less"}

\item{p.cutoff}{Default 0.05}

\item{difference.cutoff}{Default 0}

\item{sep}{Default ", "}
}
\value{

}
\description{
Iterate all the combination of housekeeping genes
}
