% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCA.R
\name{clincial_impact}
\alias{clincial_impact}
\title{Plot clinical impact: detail number of patients}
\usage{
clincial_impact(risk.score, label, name = "Study", palette = "aaas")
}
\arguments{
\item{name}{Default "Study"}

\item{palette}{Default "aaas"}
}
\description{
Plot clinical impact: detail number of patients
}
