% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE.R
\name{ClusterProfiler.GSEA.Compare}
\alias{ClusterProfiler.GSEA.Compare}
\title{Perform GSE analysis across multiple group by cluster profiler}
\usage{
ClusterProfiler.GSEA.Compare(
  gene,
  minGSSize = 10,
  qvalue = 0.05,
  GO = FALSE,
  KEGG = FALSE,
  MSigDb = TRUE,
  Hallmark = TRUE
)
}
\arguments{
\item{gene}{A list including multiple group. Must vector names are gene symbol and vector value are log fold chage}

\item{minGSSize}{10}

\item{qvalue}{0.05}

\item{GO}{TRUE}

\item{KEGG}{TRUE}

\item{MSigDb}{TRUE}

\item{Hallmark}{TRUE}
}
\value{

}
\description{
Perform GSE analysis across multiple group by cluster profiler
}
