% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE.R
\name{ClusterProfiler.GSEA.ORA.customGS.Compare}
\alias{ClusterProfiler.GSEA.ORA.customGS.Compare}
\title{Perform GSE analysis across multiple group by cluster profiler and user customed gene sets}
\usage{
ClusterProfiler.GSEA.ORA.customGS.Compare(
  gene,
  customGS = NULL,
  minGSSize = 10,
  qvalue = 0.05,
  gse = FALSE,
  ova = FALSE
)
}
\arguments{
\item{customGS}{ClusterProfiler::read.gmt()  term to symbols}

\item{minGSSize}{10}

\item{qvalue}{0.05}

\item{gse}{Default FALSE, gene set enrichment analysis}

\item{ova}{Default FALSE, overrepresentation test}
}
\description{
Perform GSE analysis across multiple group by cluster profiler and user customed gene sets
}
