% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE.R
\name{ClusterProfiler.OverRepresentationTest.Compare}
\alias{ClusterProfiler.OverRepresentationTest.Compare}
\title{Perform over-representation test across multiple group by cluster profiler}
\usage{
ClusterProfiler.OverRepresentationTest.Compare(
  gene,
  minGSSize = 10,
  qvalue = 0.05,
  GO = FALSE,
  KEGG = FALSE,
  MSigDb = TRUE,
  Hallmark = TRUE,
  exp.gene.type = "ENSEMBL"
)
}
\arguments{
\item{gene}{A list including multiple group. Must gene symbol}

\item{minGSSize}{10}

\item{qvalue}{0.05}

\item{GO}{TRUE}

\item{KEGG}{TRUE}

\item{MSigDb}{TRUE}

\item{Hallmark}{TRUE}

\item{exp.gene.type}{RNA expression ID ENSEMBL. keytypes(org.Hs.eg.db)}
}
\description{
Perform over-representation test across multiple group by cluster profiler
}
