% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE.R
\name{ClusterProfiler.OverRepresentationTest}
\alias{ClusterProfiler.OverRepresentationTest}
\title{Perform over-representation test by cluster profiler}
\usage{
ClusterProfiler.OverRepresentationTest(
  gene,
  minGSSize = 10,
  qvalue = 0.05,
  GO = TRUE,
  KEGG = TRUE,
  MSigDb = TRUE,
  Hallmark = TRUE
)
}
\arguments{
\item{gene}{A vector. Must gene symbol}

\item{minGSSize}{10}

\item{qvalue}{0.05}

\item{GO}{TRUE}

\item{KEGG}{TRUE}

\item{MSigDb}{TRUE}

\item{Hallmark}{TRUE}
}
\value{

}
\description{
Perform over-representation test by cluster profiler
}
