% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{compare.differential.analysis}
\alias{compare.differential.analysis}
\title{Perform differential analysis (1 vs others)}
\usage{
compare.differential.analysis(
  rna.df.log,
  group,
  prefix = "Group",
  cal.auc = FALSE
)
}
\arguments{
\item{prefix}{Default "Group"}

\item{cal.auc}{If calculate AUC value}
}
\value{

}
\description{
Perform differential analysis (1 vs others)
}
