% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE.R
\name{compare.GSE.HTSAnalyzer}
\alias{compare.GSE.HTSAnalyzer}
\title{Perform GSEA across multiple groups. Compare 1 vs other}
\usage{
compare.GSE.HTSAnalyzer(
  rna.df.log,
  group,
  prefix = "Group",
  customGS = NULL,
  exp.gene.type = "ENSEMBL",
  cutoff.log10 = 4
)
}
\arguments{
\item{prefix}{Default "Group"}

\item{customGS}{User customed gene set. qusage::read.gmt. Should be converted to ENTREZID}

\item{exp.gene.type}{RNA expression ID. keytypes(org.Hs.eg.db)}

\item{cutoff.log10}{Default 4. Minimux or maximum log10 value. Useful when meet inf or draw heatmap}
}
\value{

}
\description{
Perform GSEA across multiple groups. Compare 1 vs other
}
\examples{
This function will perform GSEA analysis. Default geneset: GOBP, C2, C5, KEGG, HALLMARK

}
