% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{compareROC}
\alias{compareROC}
\title{Compare two ROC curves}
\usage{
compareROC(
  risk1,
  lab1,
  risk2,
  lab2,
  method = c("delong", "bootstrap", "venkatraman"),
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{method}{the method to use, either “delong”, “bootstrap” or “venkatraman”. The first letter is sufficient.}

\item{alternative}{"two.sided", "less", "greater"}
}
\description{
Compare two ROC curves
}
