% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{confidence_interval}
\alias{confidence_interval}
\title{confidence interval}
\usage{
confidence_interval(vector, interval)
}
\arguments{
\item{interval}{}
}
\value{

}
\description{
confidence interval
}
\examples{
https://stackoverflow.com/questions/48612153/how-to-calculate-confidence-intervals-for-a-vector

vector <- c(12, 17, 24, 35, 23, 34, 56)
confidence_interval(vector, 0.90)
}
