% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Get confusion matrix}
\usage{
confusion_matrix(groups, risk.pro, cancer = "Cancer", best.cutoff = NA)
}
\arguments{
\item{groups}{TRUE/FALSE label}

\item{cancer}{Cancer Name}

\item{best.cutoff}{If not set, use youden index instead}

\item{rs}{Predicted score}
}
\value{
A data.frame object
}
\description{
Get confusion matrix
}
\examples{
label = c(1,1,1,1,1,2,2,2,2,2) == 1
risk.probability = runif(10, min=0, max=100)
confusion_matrix(label, risk.probability, cancer = "ESCC")
}
