% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtype.R
\name{consensusNodeJaccard}
\alias{consensusNodeJaccard}
\title{Calculate Jaccard similarity coefficient}
\usage{
consensusNodeJaccard(df, returnDf = T, concateStudy = F)
}
\arguments{
\item{df}{row is sample, column is study subtype}

\item{returnDf}{reture a n*n matrix}

\item{concateStudy}{}
}
\value{

}
\description{
Calculate Jaccard similarity coefficient
}
