% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtyping.R
\name{consensusSubtyping}
\alias{consensusSubtyping}
\title{Identification of consensus subtypes}
\usage{
consensusSubtyping(
  df,
  replicate = 100,
  seed = 1,
  proportion = 0.8,
  adjusted.hypergeometrix.p = F,
  inflation = 2,
  adjacencyMatrixCutoff = NULL,
  subtype.prefix = ""
)
}
\arguments{
\item{df}{row is sample, column is study subtype. Plsease note rownames is sample ID}

\item{replicate}{Default 100}

\item{seed}{Default 1}

\item{proportion}{Default 0.8}

\item{adjusted.hypergeometrix.p}{Default 0.05}

\item{inflation}{Default 2}

\item{adjacencyMatrixCutoff}{Default NULL. Cutoff for adjacenty matrix.}

\item{subtype.prefix}{}
}
\value{

}
\description{
Identification of consensus subtypes
}
\examples{
data("Subtype.matrix")
res = loonR::consensusSubtyping(Subtype.matrix, replicate = 50)
}
