% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{convertDfToNumeric}
\alias{convertDfToNumeric}
\title{Convert data frame to numeric}
\usage{
convertDfToNumeric(df)
}
\arguments{
\item{df}{}
}
\value{
A data.frame with numeric element in it
}
\description{
Convert data frame to numeric
}
\examples{
convertDfToNumeric(df)
}
