% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDMapping.R
\name{convertEnsemblDF2SymbolDF}
\alias{convertEnsemblDF2SymbolDF}
\title{Convert Ensembl data frame to gene symbol dataframe}
\usage{
convertEnsemblDF2SymbolDF(ensembl.df)
}
\arguments{
\item{ensembl.df}{Row is ENSG gene, col is sample}
}
\description{
Convert Ensembl data frame to gene symbol dataframe
}
