% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immune.R
\name{CorrelationHeatmapTwoDf}
\alias{CorrelationHeatmapTwoDf}
\title{Perform correlation analysis between two data.frame}
\usage{
CorrelationHeatmapTwoDf(x.axis.df, y.axis.df, cor.method = "spearman")
}
\arguments{
\item{x.axis.df}{Row is sample, column variable will be column names in new correlation heatmap.}

\item{y.axis.df}{Row is sample, column variable will be row names in new correlation heatmap.}

\item{cor.method}{Defaut spearman}
}
\value{
A ggplot2 object
}
\description{
Perform correlation analysis between two data.frame
}
