% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immune.R
\name{CorrelationHeatmapTwoDfMantelTest}
\alias{CorrelationHeatmapTwoDfMantelTest}
\title{Perform correlation analysis between two data.frame}
\usage{
CorrelationHeatmapTwoDfMantelTest(
  x.axis.df,
  y.axis.df,
  cor.method = "spearman",
  spec.select.list = NULL,
  rho.cutoff = 0.3,
  p.color = c("#D95F02", "#CCCCCC99", "#1B9E77")
)
}
\arguments{
\item{x.axis.df}{Row is sample, column variable will be column names in new correlation heatmap.}

\item{y.axis.df}{Row is sample, column variable will be row names in new correlation heatmap.}

\item{cor.method}{Defaut spearman}

\item{spec.select.list}{list(group1 = c(1,2,3), grou2 = c(4,5)) numbers are col indexin x.axis.df}

\item{rho.cutoff}{Default 0.3}

\item{p.color}{c("#D95F02", "#CCCCCC99", "#1B9E77")}
}
\value{
A ggplot2 object
}
\description{
Perform correlation analysis between two data.frame
}
\examples{

library(vegan)
data(varespec)
data(varechem)

loonR::CorrelationHeatmapTwoDfMantelTest(varespec[,c(1,2)], varechem)

l = list(Community_Diversity= 1:5, Community_structure =6:14)
loonR::CorrelationHeatmapTwoDfMantelTest(varespec, varechem, spec.select.list = l)

}
