% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{count_to_tpm}
\alias{count_to_tpm}
\title{Convert count to TPM}
\usage{
count_to_tpm(mat, gene_length)
}
\arguments{
\item{mat}{row is gene, col is sample}

\item{gene_length}{}
}
\description{
Convert count to TPM
}
