% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtyping.R
\name{countClassByColumn}
\alias{countClassByColumn}
\title{Count the class by each column}
\usage{
countClassByColumn(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Count the class by each column
}
\examples{
data(Subtype.matrix)
loonR::countClassByColumn(Subtype.matrix)
}
