% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{cross.validation}
\alias{cross.validation}
\title{Title}
\usage{
cross.validation(df = "", label = "", k = 5, n = 100)
}
\arguments{
\item{df}{data.frame. Col is sample, row is gene}

\item{k}{Folds}

\item{n}{Repeat times}
}
\value{

}
\description{
Title
}
