% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtyping.R
\name{dataframe2gct_file}
\alias{dataframe2gct_file}
\title{Convert data frame to GenePattern gct file}
\usage{
dataframe2gct_file(
  df,
  file = "./expression.gcf",
  description = NULL,
  version = "#1.2"
)
}
\arguments{
\item{df}{Row is gene}

\item{file}{Default ./expression.gcf}
}
\value{

}
\description{
Convert data frame to GenePattern gct file
}
