% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCA.R
\name{decisionCurve.diff.pvalue}
\alias{decisionCurve.diff.pvalue}
\title{Comparing two decision curves and reporting P value}
\usage{
decisionCurve.diff.pvalue(
  outcome,
  pred1,
  pred2,
  boots = 100,
  xstart = 0.01,
  xstop = 0.99,
  step = 0.01,
  type = "treated",
  model.name = "Model"
)
}
\arguments{
\item{boots}{100 Bootstrape to avoid overfitting}

\item{xstart}{the starting point of the threshold probability, the default value is 0.01.}

\item{xstop}{the end point of the threshold probability, the default value is 0.99}

\item{step}{a numerical value specifying the incremental step of the threshold probability, the default value is 0.01}

\item{type}{controls the type of net benefit to be computed. The allowed values correspond to the treated (“treated”), untreated (“untreated”) and overall (“overall”) patients, or to the ADAPT index (“adapt”). The default is the “treated”}

\item{model.name}{Default Model}
}
\value{

}
\description{
Comparing two decision curves and reporting P value
}
\examples{
data(LIRI)

pred1 <- as.vector( unlist( LIRI[,c(3)] ) )
pred2 <- as.vector( unlist( LIRI[,c(2)] ) )
decisionCurve.diff.pvalue(LIRI$status, pred1, pred2)

}
