% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{decisionCurveAnalysis}
\alias{decisionCurveAnalysis}
\title{Decision curve analysis}
\usage{
decisionCurveAnalysis(
  data.frame.list = NULL,
  label = NULL,
  rms = FALSE,
  validation.df = NULL,
  palette = "aaas"
)
}
\arguments{
\item{data.frame.list}{row is sample. List should inlcude names}

\item{rms}{Default FALSE}

\item{validation.df}{shold include all variable in data.frame.list}
}
\value{

}
\description{
Decision curve analysis
}
\examples{
data(LIRI)

d1 <- LIRI[,c(3)]
d2 <- LIRI[,c(3,4)]
d3 <- LIRI[,c(3,4,5)]
data.frame.list = list(d1=d1,d2=d2,d3=d3)
decisionCurveAnalysis(data.frame.list, label=LIRI$status)
}
