% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{DESeq2_differential}
\alias{DESeq2_differential}
\title{Differential analysis by DESEQ2}
\usage{
DESeq2_differential(
  rawcount,
  group,
  pre.filter = 0,
  return.normalized.df = FALSE,
  cal.AUC = TRUE
)
}
\arguments{
\item{rawcount}{Only support raw count. Please note}

\item{group}{factor, first control then experiment}

\item{pre.filter}{if filter low expression gene}
}
\value{

}
\description{
Differential analysis by DESEQ2
}
