% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{determineClusterNumber}
\alias{determineClusterNumber}
\title{Determine the best cluster numbers}
\usage{
determineClusterNumber(
  df,
  distance = "euclidean",
  method = "kmeans",
  dissimilarity = NULL,
  min.nc = 2,
  max.nc = 8
)
}
\arguments{
\item{df}{Row is sample and column is variable}

\item{distance}{This must be one of: "euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski" or "NULL".}

\item{method}{kmeans. the cluster analysis method to be used. This should be one of: "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid", "kmeans".}

\item{dissimilarity}{dissimilarity matrix to be used. By default, diss=NULL, but if it is replaced by a dissimilarity matrix, distance should be "NULL".}

\item{min.nc}{2 minimal number of clusters}

\item{max.nc}{8 maximal number of clusters, between 2 and (number of objects - 1), greater or equal to min.nc.}
}
\value{

}
\description{
Determine the best cluster numbers
}
