% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{dimPlotWithCountBar}
\alias{dimPlotWithCountBar}
\title{Plot dim plot and cell count bar}
\usage{
dimPlotWithCountBar(
  object = NULL,
  color = NULL,
  dim = "umap",
  rm.axis = FALSE,
  cell.id = NULL,
  bar.width = 0.2,
  point.size = 1,
  rm.barplot = FALSE,
  legend.psize = 1.5,
  arrow.len = 0.2
)
}
\arguments{
\item{arrow.len}{}
}
\description{
Plot dim plot and cell count bar
}
\examples{
# modified from scRNAtoolVis::scatterCellPlot

}
