% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{download.geo.dataset}
\alias{download.geo.dataset}
\title{Download dataset from GEO by accession ID and platform}
\usage{
download.geo.dataset(geo.accession.id, platform, destdir = tempdir())
}
\arguments{
\item{geo.accession.id}{GEO Accession ID}

\item{platform}{Platform}

\item{destdir}{Default tempdir()}
}
\value{
list(expression, phenotype, probe.annotation)
}
\description{
Download dataset from GEO by accession ID and platform
}
