% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\name{download.TCGA.Biolinks}
\alias{download.TCGA.Biolinks}
\title{Download TCGA data through TCGABiolinks}
\usage{
download.TCGA.Biolinks(
  project.id = NULL,
  mRNA = T,
  miRNA = T,
  met = T,
  protein = T,
  mutation = T,
  cnv = T,
  clin = T,
  subtype = T,
  dir = "./"
)
}
\arguments{
\item{mRNA}{Default T}

\item{miRNA}{Default T}

\item{met}{Default T}

\item{protein}{Default T}

\item{mutation}{Default T}

\item{cnv}{Default T}

\item{clin}{Default T}

\item{dir}{Default current directory}
}
\description{
Download TCGA data through TCGABiolinks
}
