% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{draw.expression.dotplot}
\alias{draw.expression.dotplot}
\title{Draw expression by dotplot
Need to be update}
\usage{
draw.expression.dotplot(
  df,
  group,
  nrow = 4,
  stat.method = "wilcox.test",
  show.stat.p = TRUE
)
}
\arguments{
\item{df}{Row is gene, col is sample}

\item{nrow}{# of row to plot expression}

\item{stat.method}{wilcox.test or t.test}
}
\value{

}
\description{
Draw expression by dotplot
Need to be update
}
\examples{
loonR::draw.expression.dotplot(candidate.combined.df, group )
}
