% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{draw_confusion_matrix}
\alias{draw_confusion_matrix}
\title{Confusion matrix visulization}
\usage{
draw_confusion_matrix(groups, risk.pro, best.cutoff = NA)
}
\arguments{
\item{groups}{True label}

\item{best.cutoff}{If not set, use youden index instead}

\item{rs}{Predicted score}
}
\value{

}
\description{
Confusion matrix visulization
}
\examples{
draw_confusion_matrix(label,risk.probability)
}
