% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WES.R
\name{DrawMutationType}
\alias{DrawMutationType}
\title{Count each event type and draw pie chart}
\usage{
DrawMutationType(
  data,
  color = "jco",
  colid = 2,
  alpha = 0.8,
  title = "",
  border = "white"
)
}
\arguments{
\item{data}{A data.frame or list object.}

\item{color}{ggsci color palette}

\item{colid}{If provide data.frame, column id need to be set}

\item{alpha}{Alpha value in plot}

\item{title}{Pie title}

\item{border}{Border color}
}
\value{
ggplot2 object
}
\description{
This function is evry similar to SuppaEventPie. Here the pie is draw based on ggplot2.
}
\examples{
DrawMutationType(res$Func.refGene, title = "# of mutation")
or DrawMutationType(res, col = 6, title = "# of mutation")
}
