% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WES.R
\name{DrawMutationTypeBarPlot}
\alias{DrawMutationTypeBarPlot}
\title{Draw bar plot}
\usage{
DrawMutationTypeBarPlot(
  data,
  color = "jco",
  alpha = 0.8,
  title = "",
  x = "Sample",
  y = "Number of mutations (SNV+INDEL)",
  position = "stack"
)
}
\arguments{
\item{data}{A data.frame. Column names for ggplot2 :x=Sample, y=as.numeric(value), fill=variable}

\item{color}{ggsci color palette}

\item{alpha}{Color alpha}

\item{title}{Titile of plot}

\item{x}{X label}

\item{y}{Y label}

\item{position}{}
}
\value{
ggplot2 object
}
\description{
Draw bar plot
}
\examples{
data <- data.frame( variable = c("exonic","exonic","exonic","intronic","intronic","intronic"),
                    Sample = c("S1","S2","S3","S1","S2","S3"),
                    value = c(1,2,3,4,5,6)
)
DrawBarPlot(data)

}
