% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{drawScatter}
\alias{drawScatter}
\title{Draw scatter plot}
\usage{
drawScatter(
  xvalue,
  yvalue,
  xlab = "X",
  ylab = "Y",
  group = NA,
  color = "jco",
  title = "",
  margin = TRUE,
  xlim = NULL,
  ylim = NULL,
  show.sample.name = FALSE,
  label = NULL,
  cor.coef = F
)
}
\arguments{
\item{xvalue}{Vector of values}

\item{yvalue}{The same length wtih x value}

\item{color}{Default jco}

\item{margin}{Default TRUE, show margin plot}

\item{show.sample.name}{Default FALSE}

\item{label}{Default NULL}

\item{cor.coef}{TRUE/FALSE to show coefficient}
}
\value{
ggplot2 object
}
\description{
Draw scatter plot
}
\examples{
loonR::drawScatter(sample.info$before_reads, sample.info$mirdeep2_mapped)
}
