% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methylation.R
\name{ELMER_RNA_Met_analysis}
\alias{ELMER_RNA_Met_analysis}
\title{RNA and methylation analysis}
\usage{
ELMER_RNA_Met_analysis(
  rna.df = NULL,
  met.df = NULL,
  group = NULL,
  genome = "hg19",
  met.platform = "450K",
  beta.dif = 0.2,
  dir.group1 = "hyper",
  specify.probes = NULL,
  TF.motif = FALSE,
  run.diff = TRUE
)
}
\arguments{
\item{rna.df}{row name should be ensembl ID}

\item{met.df}{row name should be probeID}

\item{group}{normal 在前, cancer 在后}

\item{genome}{hg19 or hg38}

\item{met.platform}{450K or EPIC}

\item{beta.dif}{Default 0.2. Beta difference}

\item{dir.group1}{It can be "hypo" which means the probes are hypomethylated in group1; "hyper" which means the probes are hypermethylated in group1}

\item{specify.probes}{Specify methylation probe instead of ELMER}

\item{TF.motif}{Default FALSE. If perform TF and motif analysi}

\item{run.diff}{If run methylation analysis}
}
\description{
RNA and methylation analysis
}
