% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{export2ppt}
\alias{export2ppt}
\title{Export figure to PPT
https://github.com/tomwenseleers/export}
\usage{
export2ppt(obj, file = "~/test.pptx", append = TRUE)
}
\arguments{
\item{obj}{ggplot object}

\item{file}{default is "~/test.pptx"}

\item{append}{Default TRUE}
}
\value{
NA
}
\description{
Export figure to PPT
https://github.com/tomwenseleers/export
}
\examples{
export2ppt(ggplot2.obj)
}
