% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{exportScenicLoom}
\alias{exportScenicLoom}
\title{Export seurat object for SCENIC analysis}
\usage{
exportScenicLoom(seurat.obj, dir = "scenic", prefix = "scenic")
}
\arguments{
\item{dir}{Directory for export object}

\item{prefix}{Default scenic. Prefix in dir}
}
\description{
Export seurat object for SCENIC analysis
}
