% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{exportTable}
\alias{exportTable}
\title{Overwrite write.table with modified default option. Export data frame to file}
\usage{
exportTable(df, file = "~/test.tsv", quote = F, sep = "\\t", row.names = F)
}
\arguments{
\item{file}{file path}

\item{quote}{Default FALSE}

\item{sep}{Default \t}
}
\value{

}
\description{
Overwrite write.table with modified default option. Export data frame to file
}
