% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.R
\name{extractSignaturePercent}
\alias{extractSignaturePercent}
\title{Estimate signature percents}
\usage{
extractSignaturePercent(
  chr = NULL,
  Start_Position = NULL,
  End_Position = NULL,
  ref_base = NULL,
  alt_base = NULL,
  Sample_Barcode = NULL,
  signatures.ref = NULL
)
}
\arguments{
\item{signatures.ref}{Signature * 96 context, colnames should be the same as signatures.nature2013 dataset}
}
\value{

}
\description{
Estimate signature percents
}
