% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{feature.selection.4givenMLAlgorithm}
\alias{feature.selection.4givenMLAlgorithm}
\title{Variable Importance from a specific Machine Learning Algorithms}
\usage{
feature.selection.4givenMLAlgorithm(
  feature.df,
  group,
  seed = 111,
  scale = TRUE,
  method = "glm"
)
}
\arguments{
\item{feature.df}{Row is sample, column is feature}

\item{scale}{}

\item{method}{glm for logistic, rf for random forest, RRF for Regularized Random Forest, rpart for rpart model. https://www.rdocumentation.org/packages/caret/versions/4.47/topics/train}
}
\value{

}
\description{
Variable Importance from a specific Machine Learning Algorithms
}
\examples{
loonR::feature.selection.4givenMLAlgorithm(miR.df, group, method = 'glm')
}
