% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{feature.selection.boruta}
\alias{feature.selection.boruta}
\title{Feature selection by Boruta}
\usage{
feature.selection.boruta(
  feature.df,
  group,
  seed = 111,
  scale = TRUE,
  withTentative = FALSE
)
}
\arguments{
\item{feature.df}{Row is sample, column is feature}

\item{seed}{Default 111}

\item{withTentative}{Default FALSE, not include tentative variable}
}
\description{
Feature selection by Boruta
}
\examples{
loonR::feature.selection.boruta(miR.df, group)
}
