% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{fillNADataframe}
\alias{fillNADataframe}
\title{Fill NA by row or column}
\usage{
fillNADataframe(df, byRow = FALSE, byColumn = FALSE, f = "mean")
}
\arguments{
\item{f}{Default "mean"}
}
\description{
Fill NA by row or column
}
\examples{
set.seed(1)
data.df = data.frame(
  S1 = rnorm(5),
  S2 = rnorm(5),
  S3 = rnorm(5)
)
data.df[1,2] = NA
data.df[3,3] = NA
loonR::fillNADataframe(data.df)
}
