% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtype.R
\name{fillSymmetricNATable}
\alias{fillSymmetricNATable}
\title{Fill NA in symmetric table}
\usage{
fillSymmetricNATable(symmetric.df)
}
\arguments{
\item{symmetric.df}{}
}
\value{

}
\description{
Fill NA in symmetric table
}
