% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{filterGenes}
\alias{filterGenes}
\title{Filter genes by kinks of criteria}
\usage{
filterGenes(
  df,
  expression.cutoff = 0,
  proportion.expressed.samples = 0.7,
  mean.expression = 0
)
}
\arguments{
\item{mean.expression}{}
}
\description{
Filter genes by kinks of criteria
}
