% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{Find_doublet}
\alias{Find_doublet}
\title{Find doublet}
\usage{
Find_doublet(
  data,
  PCs = 1:20,
  sct = FALSE,
  Doubletrate = 0.05,
  base_cluster = "RNA_snn_res.0.1"
)
}
\arguments{
\item{Doubletrate}{}
}
\description{
Find doublet
}
