% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{find_outlier_index}
\alias{find_outlier_index}
\title{Find outlier position}
\usage{
find_outlier_index(x, zscore = NULL, quantile = NULL)
}
\arguments{
\item{quantile}{}
}
\description{
Find outlier position
}
