% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{findMaxMinColumnNamesForEachRow}
\alias{findMaxMinColumnNamesForEachRow}
\title{Find the column name of maximum/minimum value for each row}
\usage{
findMaxMinColumnNamesForEachRow(
  df,
  max = FALSE,
  min = FALSE,
  ties.method = c("random", "first", "last"),
  specified.column = NULL
)
}
\arguments{
\item{df}{A dataframe}

\item{max}{Default FALSE}

\item{min}{Default FALSE}

\item{ties.method}{"random", "first", "last"}

\item{specified.column}{Numeric vector. Which columns want to be checked}
}
\description{
Find the column name of maximum/minimum value for each row
}
\examples{
data(LIRI)
df = LIRI
res <- loonR::findMaxMinColumnNamesForEachRow(df, min = T, specified.column = 3:6)
res
}
