% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{findRepeatedGenes}
\alias{findRepeatedGenes}
\title{Return a vector indicate gene overlapping count}
\usage{
findRepeatedGenes(
  geneColumnName = NULL,
  pColumnName = NULL,
  p.cutoff = 0.05,
  ...
)
}
\arguments{
\item{geneColumnName}{colunm name for find overlapping genes}

\item{pColumnName}{column name of p value}

\item{p.cutoff}{p value cutoff, default 0.05}

\item{...}{}
}
\description{
Return a vector indicate gene overlapping count
}
