% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.R
\name{findSomaticSignature}
\alias{findSomaticSignature}
\title{Identify signature from sample data}
\usage{
findSomaticSignature(
  chr = NULL,
  Start_Position = NULL,
  End_Position = NULL,
  hg = 38,
  pca = FALSE,
  nmf = FALSE,
  ref_base = NULL,
  alt_base = NULL,
  Sample_Barcode = NULL,
  project_name = "Project",
  n_sigs = 5
)
}
\arguments{
\item{n_sigs}{the number of signature will be identified}
}
\value{

}
\description{
Identify signature from sample data
}
