% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{formatTable}
\alias{formatTable}
\title{Format by digit}
\usage{
formatTable(df, digit = 2)
}
\arguments{
\item{df}{numeric data.frame}

\item{digit}{default = 2}
}
\value{

}
\description{
Format by digit
}
\examples{
If you want to format by column respectively
tab_test <- structure(list(col1 = c(2e-05, 0.001, 0.568978554112256), col2 = c(3.5,45.6546548788, 12585.5663), col3 = c(200, 34.52, NA)), class = "data.frame", row.names = c(NA,-3L))
as.data.frame(Map(loonR::formatTable, tab_test, c(3L, 2L, 2L)))
}
