% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{gapStat}
\alias{gapStat}
\title{Gap statistics}
\usage{
gapStat(df, dist = "spearman", method = "average")
}
\arguments{
\item{df}{Row is sample, column is variable}

\item{dist}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman"}

\item{method}{Default "Average". The agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC)}
}
\value{

}
\description{
Gap statistics
}
