% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorRelated.R
\name{generate.sequencial.color}
\alias{generate.sequencial.color}
\title{Generate sequencial color}
\usage{
generate.sequencial.color(colors = c("#c1e7ff", "#f3babc"), length = 10)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
Generate sequencial color
}
