% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{generateCombinations}
\alias{generateCombinations}
\title{Gnenerate combinations with specified size}
\usage{
generateCombinations(
  panel = NULL,
  size = 0,
  repeats = FALSE,
  vector = FALSE,
  sep = "-"
)
}
\arguments{
\item{panel}{panel which will be used to make combinations}

\item{size}{size of a single combination}

\item{repeats}{Default FALSE, not allow repeats}

\item{sep}{If return a vector, sep is needed}
}
\value{

}
\description{
Gnenerate combinations with specified size
}
