% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{generateGenePairValueDf}
\alias{generateGenePairValueDf}
\title{Generate gene pair value data.frame}
\usage{
generateGenePairValueDf(df, upRegulatedGene = NULL)
}
\arguments{
\item{df}{row is gene, column is sample}

\item{upRegulatedGene}{If input upRegulatedGene, ratio and rank will consider high/low.}
}
\description{
Generate gene pair value data.frame
}
\examples{

set.seed(1)
df = matrix(rnorm(100*10, 1, .5), ncol=10)
colnames(df) = paste0("ID",1:10)
rownames(df) = paste0("Gene",1:100)
res = loonR::generateGenePairValueDf(df)
head(res$Binary.res)
}
