% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{generatePairValue}
\alias{generatePairValue}
\title{Generate gene pair value}
\usage{
generatePairValue(value = NULL, names = NULL, sep = "_")
}
\arguments{
\item{sep}{Pair seperation}
}
\description{
Generate gene pair value
}
\examples{

set.seed(111)
v = rnorm(10)
n = paste0("ID", 1:10)
names(v) = n
loonR::generatePairValue(v,n)

}
