% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{genereateLabelsByGroup}
\alias{genereateLabelsByGroup}
\title{Generate label based on group}
\usage{
genereateLabelsByGroup(label = NULL, ...)
}
\arguments{
\item{label}{Number of elements in label vector should be the same with total group number}

\item{...}{one or more group vector}
}
\description{
Generate label based on group
}
\examples{
n.group = c("N1", "N2", "N3", "N4")
t.group = c("T1", "T2")
m.group = c("M1", "M2", "M1")

labels <- genereateLabelsByGroup(c("N","T", "M"), n.group, t.group, m.group)
table(labels)

}
