% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{get.all.point.ROC}
\alias{get.all.point.ROC}
\title{Get all point if you want to draw yourself}
\usage{
get.all.point.ROC(pred, label, ci = FALSE, boot.n = 2000)
}
\arguments{
\item{ci}{if CI}

\item{boot.n}{2000}
}
\description{
Get all point if you want to draw yourself
}
