% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorRelated.R
\name{get.ggsci.color}
\alias{get.ggsci.color}
\title{Get colors by ggsci}
\usage{
get.ggsci.color(palette = "nrc", n = 7, alpha = 1)
}
\arguments{
\item{palette}{default: nrc. Palette ref: https://nanx.me/ggsci/articles/ggsci.html}

\item{n}{number of colors, default n = 7}

\item{alpha}{1}
}
\description{
Get colors by ggsci
}
\examples{
get.ggsci.color("nrc", n = 2, alpha = 0.7)
}
