% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA.R
\name{get.miRNAExpression}
\alias{get.miRNAExpression}
\title{Get TCGA miRNA expression data}
\usage{
get.miRNAExpression(
  tcga.project = NULL,
  rawCount = FALSE,
  CPM = FALSE,
  log2 = FALSE
)
}
\arguments{
\item{tcga.project}{See ?RTCGA.miRNASeq::miRNASeq. For example: COAD}

\item{rawCount}{If to download raw count data}

\item{CPM}{If to download CPM data}

\item{log2}{If to perform log2 transformation}
}
\value{

}
\description{
Get TCGA miRNA expression data
}
