% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorRelated.R
\name{get.mostDistint.color.palette}
\alias{get.mostDistint.color.palette}
\title{Generate a most distinctive color palatte}
\usage{
get.mostDistint.color.palette(n = 20, seed = 123)
}
\arguments{
\item{n}{Number of colors}

\item{seed}{123}
}
\value{
Color vector
}
\description{
Generate a most distinctive color palatte
}
