% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorRelated.R
\name{get.palette.color}
\alias{get.palette.color}
\title{Use ggpubr function to get all kinds of color palette
RColorBrewer, ggsci, hue or grey/gray color palettes}
\usage{
get.palette.color(palette = "nrc", n = 7, alpha = 1)
}
\arguments{
\item{palette}{Default npg. The color palette to be used for coloring or filling by groups.
Allowed values include "grey" for grey color palettes; brewer palettes e.g. "RdBu", "Blues", ...;
or custom color palette e.g. c("blue", "red");
and scientific journal palettes from ggsci R package, e.g.: "npg", "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".}

\item{n}{number of colors, default n = 7}

\item{alpha}{default 1}
}
\value{

}
\description{
Use ggpubr function to get all kinds of color palette
RColorBrewer, ggsci, hue or grey/gray color palettes
}
\examples{
get.palette.color("nrc", n = 2, alpha = 0.7)
}
