% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\name{get.TCGA.Project.Clinical}
\alias{get.TCGA.Project.Clinical}
\title{Get TCGA project clinical information}
\usage{
get.TCGA.Project.Clinical(project, subtype = FASLE, microsatellite = FALSE)
}
\arguments{
\item{project}{ESCA. TCGAbiolinks::getGDCprojects()$project_id without "TCGA_"}

\item{subtype}{if download subtype information}

\item{microsatellite}{if download microsatellite data}
}
\value{

}
\description{
Get TCGA project clinical information
}
