% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{get.TCGA.RNA.FPKM}
\alias{get.TCGA.RNA.FPKM}
\title{Get TCGA RNA expression (RPKM) data}
\usage{
get.TCGA.RNA.FPKM(tcga.project = NULL, rawCount = FALSE, log2 = FALSE)
}
\arguments{
\item{tcga.project}{See ?RTCGA.miRNASeq::miRNASeq. For example: COAD}

\item{log2}{If to perform log2 transformation}
}
\value{

}
\description{
Get TCGA RNA expression (RPKM) data
}
\examples{
This function is similar to get.TCGA.miRNAExpression in Biomarker.R file.
Data source is illumina hiseq Level 3 RSEM normalized expression data. Data from 2015-11-01 snapshot.
The RNAseq gene expression level 3 data contains Reads per Kilobase per Million mapped reads (RPKM)


}
