% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\name{get.TCGA.sample.project.infomap}
\alias{get.TCGA.sample.project.infomap}
\title{Get sample barcode and TCGA project map information}
\usage{
get.TCGA.sample.project.infomap(project = NA)
}
\arguments{
\item{project}{A vector. TCGA-ACC, TCGA-COAD. See TCGAbiolinks::getGDCprojects()$project_id}
}
\value{
A data.frame
X1 proj samples
1	1	TCGA-ACC	TCGA-OR-A5JZ
2	1	TCGA-ACC	TCGA-OR-A5JK
}
\description{
Get sample barcode and TCGA project map information
}
