% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{get.YoudenIndex}
\alias{get.YoudenIndex}
\title{Get Youden index value}
\usage{
get.YoudenIndex(pred, label)
}
\arguments{
\item{label}{}
}
\value{

}
\description{
Get Youden index value
}
\examples{
get.YoudenIndex(risk.scores, labels)
}
