% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{get_all_performance_table}
\alias{get_all_performance_table}
\title{Return a table including all metrics across different threshold}
\usage{
get_all_performance_table(score = NULL, label = NULL)
}
\arguments{
\item{label}{}
}
\description{
Return a table including all metrics across different threshold
}
\examples{
score = c(10:20)
label = c(0,0,1,0,1,0,1,1,0,1,1)
loonR::get_all_performance_table(score, label)
}
