% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDMapping.R
\name{get_full_mapping_table}
\alias{get_full_mapping_table}
\title{Retrieve full ID mapping table}
\usage{
get_full_mapping_table()
}
\value{
A full annotation table
}
\description{
Retrieve full ID mapping table
}
\examples{
get_full_mapping_table
}
