% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{get_individual_candidates_performance}
\alias{get_individual_candidates_performance}
\title{Get individual_candidates_performance}
\usage{
get_individual_candidates_performance(
  scores,
  labels,
  best.cutoff = NA,
  digit = 2
)
}
\arguments{
\item{scores}{Df or list, if data.frame, column is feature.}

\item{labels}{Vector or list}

\item{digit}{2 for 0.01, 3 for 0.001}
}
\value{

}
\description{
Get individual_candidates_performance
}
