% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptsRelated.R
\name{getFastaSummary}
\alias{getFastaSummary}
\title{Obtain sequence length from fasta file}
\usage{
getFastaSummary(fa.file, cores = 20)
}
\arguments{
\item{cores}{Default 20. Speed up}
}
\value{
A data.frame object
}
\description{
Obtain sequence length from fasta file
}
\examples{
getFastaSummary(fa.path)
}
