% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptsRelated.R
\name{getGeneRelatedTranscripts.Gff}
\alias{getGeneRelatedTranscripts.Gff}
\title{Retrieve all transcripts related to target genes from gtf/gff file}
\usage{
getGeneRelatedTranscripts.Gff(
  gff.Path,
  gene.vector,
  max.transcript.length = 1500000
)
}
\arguments{
\item{gff.Path}{Annotation file path or gff GRanges object}

\item{gene.vector}{Target genes}

\item{max.transcript.length}{Transcript start - end, default 1500000}
}
\value{
plyranges object
}
\description{
Retrieve all transcripts related to target genes from gtf/gff file
}
