% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorRelated.R
\name{GetMostDistinctColorPalette}
\alias{GetMostDistinctColorPalette}
\title{Generate a most distinctive color palatte}
\usage{
GetMostDistinctColorPalette(n = 20)
}
\arguments{
\item{n}{Number of colors}
}
\value{
Color vector
}
\description{
Generate a most distinctive color palatte
}
