% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptsRelated.R
\name{getTranscriptLength}
\alias{getTranscriptLength}
\title{Obtain transcript length from gff file or GRanges object}
\usage{
getTranscriptLength(
  gff,
  returnTranscriptLength = FALSE,
  returnEndStartLength = FALSE
)
}
\arguments{
\item{gff}{Path or GRanges object}

\item{returnTranscriptLength}{Real transcripted length}

\item{returnEndStartLength}{Not the real transcript length. Transcript end - transcript start}
}
\value{

}
\description{
Obtain transcript length from gff file or GRanges object
}
