% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptsRelated.R
\name{getTranscriptProteinSequence}
\alias{getTranscriptProteinSequence}
\title{Retrieve protein sequence by biomaRt}
\usage{
getTranscriptProteinSequence(IDs, type = "ens_hs_transcript")
}
\arguments{
\item{type}{Default: ens_hs_transcript. Available: Possible filters are given by the listFilters function. biomaRt::listFilters(ensembl)}
}
\value{

}
\description{
Retrieve protein sequence by biomaRt
}
