% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtyping.R
\name{group2class_file}
\alias{group2class_file}
\title{Convert a vector to GenePattern class file}
\usage{
group2class_file(group, file = "./class.cls")
}
\arguments{
\item{group}{A vector}

\item{file}{Default ./class.cls}
}
\value{

}
\description{
Convert a vector to GenePattern class file
}
