% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{groupCorrelation}
\alias{groupCorrelation}
\title{Correlation plot based on average expression}
\usage{
groupCorrelation(seurat.obj, group.by = "ident", features = NULL)
}
\arguments{
\item{features}{}
}
\value{
correlation plot
}
\description{
Correlation plot based on average expression
}
