% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{heatmap.with.lgfold.riskpro}
\alias{heatmap.with.lgfold.riskpro}
\title{Plot heatmap with log 2 fold change and risk probability information}
\usage{
heatmap.with.lgfold.riskpro(
  heatmap.df,
  label,
  risk.pro,
  lgfold = NA,
  scale = TRUE,
  group.name = "Cancer",
  bar.name = "Log2FC",
  ylim = c(0, 1),
  show.lgfold = TRUE,
  show.risk.pro = TRUE,
  height = 5,
  show_column_names = FALSE,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  z.score.cutoff = 2,
  specified.color = c("#0c3e74", "#77a8cd", "white", "#d86652", "#7e0821")
)
}
\arguments{
\item{heatmap.df}{Row: miRNA, Column: Sample}

\item{label}{A factor with labels TURE/FALSE}

\item{lgfold}{corresponded to row name sequence}

\item{scale}{Default "TRUE"}

\item{group.name}{Default "Cancer"}

\item{ylim}{= c(0, 1), risk score range.}

\item{show.lgfold}{= TRUE。 Whether to show right panel.}

\item{show_column_names}{Default False}

\item{z.score.cutoff}{Default 2}

\item{specified.color}{Default c("#0c3e74","#77a8cd","white","#d86652","#7e0821") or colorRampPalette(c("navy", "white", "firebrick3"))(50)}
}
\value{
A heatmap plot by complex heatmap
}
\description{
Plot heatmap with log 2 fold change and risk probability information
}
\examples{
heatmap.with.lgfold.riskpro(data.tmp[candi,],label, logfd,  risk.pro)
}
