% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCA.R
\name{HosmerLemesGoodnessOfFitTest}
\alias{HosmerLemesGoodnessOfFitTest}
\title{Hosmer-Lemeshow Goodness of Fit (GOF) Test}
\usage{
HosmerLemesGoodnessOfFitTest(pred = NULL, label = NULL, bin = 10)
}
\arguments{
\item{pred}{pred score}

\item{label}{Label}

\item{bin}{Default 10}
}
\value{

}
\description{
Hosmer-Lemeshow Goodness of Fit (GOF) Test
}
\examples{
set.seed(123)
}
