% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{hyperGeoTest}
\alias{hyperGeoTest}
\title{Title}
\usage{
hyperGeoTest(
  row.group,
  col.group,
  row.prefix = "",
  col.prefix = "",
  lower.tail = FALSE
)
}
\arguments{
\item{lower.tail}{Default FALSE}
}
\value{

}
\description{
Title
}
