% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtyping.R
\name{hyperGeoTest4ContingencyTable}
\alias{hyperGeoTest4ContingencyTable}
\title{Hyper geometrix test for contigency table}
\usage{
hyperGeoTest4ContingencyTable(df, lower.tail = FALSE)
}
\arguments{
\item{df}{data.frame with row and col names, and numeric value}

\item{lower.tail}{}
}
\description{
Hyper geometrix test for contigency table
}
\examples{
df = data.frame(a=c(1,4),b=c(2,19))
rownames(df) = c("c","d")
loonR::hyperGeoTest4ContingencyTable(df)
}
