% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDMapping.R
\name{id_mapping}
\alias{id_mapping}
\title{Title}
\usage{
id_mapping(IDS, key = "ENSEMBL", column = c("SYMBOL"), clean = FALSE)
}
\arguments{
\item{clean}{If set TRUE, will only return mapped ID without duplicates}
}
\description{
Title
}
\examples{
keytypes(org.Hs.eg.db)
 [1] "ACCNUM"       "ALIAS"        "ENSEMBL"      "ENSEMBLPROT"  "ENSEMBLTRANS" "ENTREZID"
 [7] "ENZYME"       "EVIDENCE"     "EVIDENCEALL"  "GENENAME"     "GO"           "GOALL"
 [13] "IPI"          "MAP"          "OMIM"         "ONTOLOGY"     "ONTOLOGYALL"  "PATH"
 [19] "PFAM"         "PMID"         "PROSITE"      "REFSEQ"       "SYMBOL"       "UCSCKG"
 [25] "UNIGENE"      "UNIPROT"
}
