% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtyping.R
\name{identifySubtypeFromMatrix}
\alias{identifySubtypeFromMatrix}
\title{A matrix}
\usage{
identifySubtypeFromMatrix(
  df,
  adjacency.cutoff = 0.5,
  inflation = 2,
  usingRawDf = F,
  clusterPrefix = NULL
)
}
\arguments{
\item{df}{n*n row and column are samples}

\item{inflation}{Default is 2 for mcl clustering}

\item{usingRawDf}{Default FALSE. If use raw data frame instead of adjacency matrix}

\item{clusterPrefix}{}
}
\value{

}
\description{
A matrix
}
\examples{
data(Subtype.matrix)
sub.asso.ana.res <- loonR::subtypeAssociationAnalysis(Subtype.matrix)
cluster.res = loonR::identifySubtypeFromMatrix(sub.asso.ana.res$JaccardAnalysis, adjacency.cutoff = 0.8)
cluster.res
}
