% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Generic is.nan function}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Generic is.nan function
}
\examples{
# This function can be applied to data.frame
# Just use is.nan(data.frame)
}
