% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{lasso.cox.selection}
\alias{lasso.cox.selection}
\title{LASSO Cox selection}
\usage{
lasso.cox.selection(d.frame, status, time, seed = 666, scale = TRUE)
}
\arguments{
\item{d.frame}{Data.frame --- Row: sample, Column: gene expression}

\item{time}{OS, DFS, RFS et al.....}

\item{seed}{Default 666}

\item{scale}{}
}
\value{

}
\description{
LASSO Cox selection
}
\examples{
data(LIRI)
loonR::lasso.cox.selection(LIRI[,3:5],LIRI$status, LIRI$time)
}
