% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{lasso.cv.select.feature}
\alias{lasso.cv.select.feature}
\title{Perform multiple round lasso to select stable feature}
\usage{
lasso.cv.select.feature(
  data.matrix,
  label,
  folds = 5,
  seed = 666,
  n = 100,
  family = "binomial",
  type.measure = "class",
  cores = 50,
  scale = TRUE
)
}
\arguments{
\item{data.matrix}{Row is sample}

\item{n}{100}

\item{family}{Default binomial}

\item{type.measure}{class, auc, deviance, mae. “deviance” uses actual deviance. “mae” uses mean absolute error. “class” gives misclassification error. “auc” (for two-class logistic regression ONLY) gives area under the ROC curve.}

\item{cores}{50}

\item{scale}{Default TRUE}
}
\value{

}
\description{
Perform multiple round lasso to select stable feature
}
