% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{limma_differential}
\alias{limma_differential}
\title{Differential analysis by LIMMA}
\usage{
limma_differential(df, group, rawcount = FALSE, voom = FALSE)
}
\arguments{
\item{df}{raw count or log2(TPM+1), default log2(TPM+1)}

\item{group}{}

\item{rawcount}{true or false}

\item{voom}{true or false. If library size changed too much.}
}
\value{

}
\description{
Differential analysis by LIMMA
}
