% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE.R
\name{load.gmt}
\alias{load.gmt}
\title{Read in gene set information from .gmt files}
\usage{
load.gmt(gmt.path)
}
\arguments{
\item{gmt.path}{Please make sure the second column is description. The name of a file to read data values from. Should be a tab-separated text file, with one row per gene set. Column 1 has gene set names (identifiers), column 2 has gene set descriptions, remaining columns are gene ids for genes in that geneset}
}
\value{
A list, where each index represents a separate gene set.
}
\description{
Read in gene set information from .gmt files
}
