% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{load.rsem.matrix}
\alias{load.rsem.matrix}
\title{Load quantification result from RSEM by tximport}
\usage{
load.rsem.matrix(dirpath, isoform = FALSE, subdirs = TRUE)
}
\arguments{
\item{dirpath}{Simple set the directory which contains RSEM output folder}

\item{isoform}{specify data type. isoforma specific or not.}

\item{subdirs}{If existing in subdirectory}
}
\value{

}
\description{
Load quantification result from RSEM by tximport
}
\examples{
Directory tree subdirs = TRUE: dir/sample/prefix.genes.results or
Directory tree subdirs = TRUE: dir/sample/prefix.isoforms.results
Directory tree subdirs = FALSE: dir/sample.prefix.genes.results or
Directory tree subdirs = FALSE: dir/sample.prefix.isoforms.results
}
