% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{load.salmon.matrix}
\alias{load.salmon.matrix}
\title{Read Salmon output}
\usage{
load.salmon.matrix(dirPath)
}
\arguments{
\item{dirPath}{Simple set the directory which contains Salmon output folder}
}
\value{
A tximport oject
}
\description{
Read Salmon output
}
\examples{
Directory tree: dir/sample_quant/quant.sf

}
