% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methylation.R
\name{loadArraryData}
\alias{loadArraryData}
\title{Title}
\usage{
loadArraryData(dir, arraytype = "EPIC")
}
\arguments{
\item{arraytype}{EPIC or 450K}
}
\value{

}
\description{
Title
}
\examples{
Reference: https://bioconductor.org/packages/release/bioc/vignettes/ChAMP/inst/doc/ChAMP.html

}
