% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{loadArrayExpressDataset}
\alias{loadArrayExpressDataset}
\title{Download dataset from ArrayExpress}
\usage{
loadArrayExpressDataset(
  accession_id = NULL,
  Processed.data = NULL,
  relationship = NULL,
  Array.design = NULL
)
}
\arguments{
\item{Processed.data}{processed.1.zip}

\item{relationship}{sdrf.txt}

\item{Array.design}{adf.txt}
}
\value{

}
\description{
Download dataset from ArrayExpress
}
\examples{
# https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6134/
accession_id = "E-MTAB-6134"
Processed.data = "E-MTAB-6134.processed.1.zip"
relationship = "E-MTAB-6134.sdrf.txt"
Array.design = "A-GEOD-13667.adf.txt"

loonR::downloadArrayExpressDataset(accession_id)

}
