% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methylation.R
\name{loadMethylationArraryData}
\alias{loadMethylationArraryData}
\title{Title}
\usage{
loadMethylationArraryData(dir, arraytype = "EPIC")
}
\arguments{
\item{arraytype}{EPIC or 450K}
}
\value{

}
\description{
Title
}
\examples{
Reference: https://bioconductor.org/packages/release/bioc/vignettes/ChAMP/inst/doc/ChAMP.html

}
