% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{log2dfQuantileNormalization}
\alias{log2dfQuantileNormalization}
\title{Quantile normalization for log2 data frame}
\usage{
log2dfQuantileNormalization(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Quantile normalization for log2 data frame
}
