% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{loo.cv.cox}
\alias{loo.cv.cox}
\title{Leave one out cross validation for cox regression}
\usage{
loo.cv.cox(df, time, status, seed = 999, label = NA)
}
\arguments{
\item{df}{Row is sample, column is gene/minRA/variat}

\item{seed}{Default 999}

\item{label}{Default NA. Recommend to provide such information}
}
\value{

}
\description{
Leave one out cross validation for cox regression
}
