% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.R
\name{maf2binaryTable}
\alias{maf2binaryTable}
\title{Convert MAF format to a binary matrix}
\usage{
maf2binaryTable(maf)
}
\arguments{
\item{maf}{File path or data.frame}
}
\description{
Convert MAF format to a binary matrix
}
