% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{MannWhitneyU_WilcoxonRankSumTest_differential}
\alias{MannWhitneyU_WilcoxonRankSumTest_differential}
\title{Calculate p value by MannWhitneyU (WilcoxonRankSum) Test manually}
\usage{
MannWhitneyU_WilcoxonRankSumTest_differential(
  df,
  group,
  cal.AUC = TRUE,
  exclude.zore = FALSE,
  exclude.na = TRUE,
  alternative = "two.sided",
  cores = 40,
  paired = FALSE
)
}
\arguments{
\item{df}{Row is gene, column is sample}

\item{alternative}{c("two.sided", "less", "greater")}

\item{paired}{Default FALSE}
}
\value{

}
\description{
Calculate p value by MannWhitneyU (WilcoxonRankSum) Test manually
}
\examples{
Nonparametric Tests of Group Differences
}
