% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{microarray_limma_differential}
\alias{microarray_limma_differential}
\title{GEO dataset analysis by limma.}
\usage{
microarray_limma_differential(
  exp,
  group,
  check.log2 = TRUE,
  normalizeBetweenArrays = TRUE
)
}
\arguments{
\item{normalizeBetweenArrays}{}
}
\value{
Normalized expression data.frame and differential analysis result
}
\description{
GEO dataset analysis by limma.
}
\examples{
# Code from GEO2E. Thanks
library(dplyr)
GSE32575 <- loonR::download.geo.dataset("GSE32575", "GPL6102")
control <- GSE32575$phenotype \%>\%
  filter(characteristics_ch1.2 == "disease state: obese before bariatric surgery") \%>\%
  pull(geo_accession)
experiment <- GSE32575$phenotype \%>\%
  filter(characteristics_ch1.2 == "disease state: obese after bariatric surgery") \%>\%
  pull(geo_accession)
exp.df <- GSE32575$expression[, c(control, experiment)]
group <- rep(c("Before", "After"), c(length(control), length(experiment)))
loonR::microarray_limma_differential(exp.df, group)
}
