% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA.R
\name{miRBaseConverter}
\alias{miRBaseConverter}
\title{miRBase convert}
\usage{
miRBaseConverter(miR.IDs = NULL, miRBase.IDs = NULL, version = NULL)
}
\arguments{
\item{miR.IDs}{A character vector representing the source miRNA names needed to be convert.}

\item{miRBase.IDs}{miRBase accessions. "MIMAT0000095"}

\item{version}{A character value representing the target miRBase version corresponding the Accessions. Default "v22"}
}
\value{

}
\description{
miRBase convert
}
