% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtyping.R
\name{multipleHyperGeoTest}
\alias{multipleHyperGeoTest}
\title{Multiple hyper geomatrix test, include 2xn, 2x2}
\usage{
multipleHyperGeoTest(
  main.row.group,
  minor.col.group,
  row.prefix = "",
  col.prefix = "",
  lower.tail = FALSE,
  title = "",
  log10.lowest = 5,
  print.fig = FALSE,
  adjusted.p = NULL,
  remove.na = FALSE,
  bar.color = "npg",
  chiq.square.test = T,
  not.consider = NA
)
}
\arguments{
\item{not.consider}{}
}
\value{

}
\description{
Multiple hyper geomatrix test, include 2xn, 2x2
}
\examples{
g1 = sample(c("G1","G2","G3"),10, replace = T)
g2 = sample(c("1","2","3"),10, replace = T)
res = loonR::hyperGeoTest(g1, g2, col.prefix = "E")
}
