% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{multiplePlotPredictedPro}
\alias{multiplePlotPredictedPro}
\title{plot predictive probability distribution}
\usage{
multiplePlotPredictedPro(group, pred, palette = "aaas", bins = 10)
}
\arguments{
\item{pred}{Must within 0-1}

\item{bins}{}
}
\value{
Contain multiple plots
}
\description{
plot predictive probability distribution
}
