% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{multivariate_or}
\alias{multivariate_or}
\title{Title Row: sample, Column: gene expression}
\usage{
multivariate_or(df, label)
}
\arguments{
\item{label}{}
}
\value{

}
\description{
Title Row: sample, Column: gene expression
}
