% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{norm_find_scale}
\alias{norm_find_scale}
\title{NormalizeData -> FindVariableFeatures -> ScaleData}
\usage{
norm_find_scale(
  seurat.obj,
  assay.name = "RNA",
  top_variable_features = 2000,
  vars.to.regress = NULL
)
}
\arguments{
\item{top_variable_features}{}
}
\description{
NormalizeData -> FindVariableFeatures -> ScaleData
}
