% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA.R
\name{normalize_miRNACount_toCPM}
\alias{normalize_miRNACount_toCPM}
\title{Normalize miRNA counts to CPM}
\usage{
normalize_miRNACount_toCPM(
  count.df,
  group,
  method = "TMM",
  log = TRUE,
  prior.count = 1
)
}
\arguments{
\item{method}{"TMM","TMMwsp","RLE","upperquartile","none"}

\item{log}{}
}
\value{

}
\description{
Normalize miRNA counts to CPM
}
