% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug.R
\name{oncoPredict}
\alias{oncoPredict}
\title{Calculate IC50 based on oncoPredict}
\usage{
oncoPredict(
  train.df = NULL,
  train.res = NULL,
  predict.df = NULL,
  minNumSamples = 10
)
}
\arguments{
\item{train.df}{path or readRDS("DataFiles/Training Data/CTRP2_Expr (TPM, not log transformed).rds")}

\item{train.res}{readRDS("DataFiles/Training Data/CTRP2_Res.rds")}

\item{minNumSamples}{}
}
\value{

}
\description{
Calculate IC50 based on oncoPredict
}
