% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{oneway.anova}
\alias{oneway.anova}
\title{Kruskal Wallis Test One Way Anova by Ranks or F test}
\usage{
oneway.anova(
  df,
  group,
  exclude.zore = FALSE,
  exclude.na = TRUE,
  cores = 40,
  parameter.test = FALSE
)
}
\arguments{
\item{df}{Row is gene, column is sample}

\item{exclude.zore}{Default FALSE}

\item{exclude.na}{Default TRUE}

\item{cores}{Default 40}

\item{parameter.test}{Default FALSE, so perform Nonparametric Tests: KruskalWallisTest}
}
\value{

}
\description{
Kruskal Wallis Test One Way Anova by Ranks or F test
}
\examples{
https://jbhender.github.io/Stats506/F18/GP/Group3.html
Nonparametric Tests: KruskalWallisTest
parametric Tests: aov  ANOVA F test
}
