% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{pairwiseAdonisTest}
\alias{pairwiseAdonisTest}
\title{pairwiseAdonis}
\usage{
pairwiseAdonisTest(data, group)
}
\arguments{
\item{data}{Row. is sample, column is feature}

\item{group}{}
}
\description{
pairwiseAdonis
}
\examples{
data(iris)
loonR::pairwiseAdonisTest(iris[,1:4], iris$Species)
}
