% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{plot.forest}
\alias{plot.forest}
\title{Forest plot}
\usage{
\method{plot}{forest}(
  tabletext,
  estimate.data,
  appendHeader = NULL,
  specify.summary = NULL,
  clipping = c(0.1, 4),
  graph.pos = "right",
  xlab = "",
  xlog = TRUE,
  xticks = c(0.1, 0.5, seq(1, 4, 1))
)
}
\arguments{
\item{tabletext}{data.frame. Column names will be used as header.}

\item{estimate.data}{Estimate, upper CI, lower CI. Corresponded with tabletext}

\item{appendHeader}{If provide a vector, top header will be added}

\item{specify.summary}{Specify the row index. Row will be bold character.}

\item{clipping}{Range to cut. Will use arrow at two cutting points}

\item{graph.pos}{Column index, or "left" or "right". Column index to put in right section}

\item{xlab}{xlab}

\item{xlog}{If in log-format. If TRUE, vectical line in 0, otherwise a vectical line at 0}

\item{xticks}{Default c(0.1, 0.5, 1, 2, 3, 4)}
}
\value{

}
\description{
Forest plot
}
\examples{
data(LIRI)
or.res <- loonR::univariate_or(LIRI[,c(1,3,4)],LIRI$status)

estimate.data = or.res[,c(2,3,4)]
text.data = data.frame(Variate = or.res$Variate, OR = or.res$OR)

plot.forest(text.data, estimate.data, graph.pos = 2, specify.summary = 1)

}
