% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{plot.pathway}
\alias{plot.pathway}
\title{Plot pathway}
\usage{
\method{plot}{pathway}(
  pathway.name = NULL,
  p = NULL,
  class = "class",
  patwahy.geneID = NULL,
  title = "Pathway enrichment"
)
}
\arguments{
\item{class}{up/dn vector}

\item{title}{}
}
\description{
Plot pathway
}
