% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{plot_half_violin}
\alias{plot_half_violin}
\title{Plot half violin}
\usage{
plot_half_violin(
  value,
  group,
  title = NULL,
  color = c("#46788E", "#78B7C9", "#F6E093", "#E58B7B", "#97B319")
)
}
\arguments{
\item{color}{}
}
\description{
Plot half violin
}
