% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{plot_miRCorrelation}
\alias{plot_miRCorrelation}
\title{Plot miR's correlation}
\usage{
plot_miRCorrelation(df, title = "miRs' correlation", cl.lim = c(-0.5, 1))
}
\arguments{
\item{df}{Row: miR expression, Column: Sample}

\item{title}{miRs' correlation}

\item{cl.lim}{c(-0.5,1) correlation lims}
}
\value{
Plot
}
\description{
Plot miR's correlation
}
\examples{
plot_miRCorrelation(data[candi,])

}
