% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{plot_pathway_bubble}
\alias{plot_pathway_bubble}
\title{Title}
\usage{
plot_pathway_bubble(
  nes.score = NULL,
  adjusted.p = NULL,
  geneset.names = NULL,
  geneset.size = NULL,
  category = NULL,
  color = c("chartreuse4", "brown2", "cornflowerblue"),
  terms.to.show = NULL
)
}
\arguments{
\item{nes.score}{required}

\item{adjusted.p}{required}

\item{geneset.names}{required}

\item{geneset.size}{required}

\item{terms.to.show}{}
}
\description{
Title
}
