% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{PlotPCA}
\alias{PlotPCA}
\title{Title}
\usage{
PlotPCA(df, group, palette = "npg")
}
\arguments{
\item{palette}{Default npg. The color palette to be used for coloring or filling by groups.
Allowed values include "grey" for grey color palettes; brewer palettes e.g. "RdBu", "Blues", ...;
or custom color palette e.g. c("blue", "red");
and scientific journal palettes from ggsci R package, e.g.: "npg", "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".}
}
\value{

}
\description{
Title
}
