% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{plotVenn}
\alias{plotVenn}
\title{Plot venn diagram}
\usage{
plotVenn(l.list, alpha = 0.5, palette = "aaas")
}
\arguments{
\item{alpha}{0.5}

\item{palette}{aaas}
}
\value{

}
\description{
Plot venn diagram
}
\examples{
l.list = list(`Up in early HCC` = early.up.genes,
`Down in early HCC` = early.down.genes,
`Up in advanced HCC` = advanced.up.genes,
`Down in advanced HCC` = advanced.down.genes)
plotVenn(l.list)
}
