% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{pre_pseudotime_matrix}
\alias{pre_pseudotime_matrix}
\title{Pseudo differential analysis}
\usage{
pre_pseudotime_matrix(mono.cds.obj, genes = NULL)
}
\arguments{
\item{genes}{}
}
\description{
Pseudo differential analysis
}
