% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{prob2logit}
\alias{prob2logit}
\title{Convert Probability to Logit}
\usage{
prob2logit(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Defined simply as \code{log(x / (1 - x))}.
}
