% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{pseudoHeatmap}
\alias{pseudoHeatmap}
\title{Pseudo heatmap}
\usage{
pseudoHeatmap(mt, genes = NULL)
}
\arguments{
\item{mt}{from pre_pseudotime_matrix function}

\item{genes}{}
}
\description{
Pseudo heatmap
}
