% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{rangeScale01}
\alias{rangeScale01}
\title{Scale the value to 0-1}
\usage{
rangeScale01(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Scale the value to 0-1
}
\examples{
s = sort(rexp(100))
loonR::range01(s)
}
