% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA.R
\name{read_mirdeep2_result}
\alias{read_mirdeep2_result}
\title{Load miRDeep2 expression matrix}
\usage{
read_mirdeep2_result(file.path)
}
\arguments{
\item{file.path}{mirdeep2 expression. Should be merged before}
}
\value{

}
\description{
Load miRDeep2 expression matrix
}
