% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA.R
\name{read_mirdeep2_resultDir}
\alias{read_mirdeep2_resultDir}
\title{Load miRDeep2 expression from a directory}
\usage{
read_mirdeep2_resultDir(dir.path, add.names2df = TRUE, message = F)
}
\arguments{
\item{dir.path}{one folder is one sample}

\item{add.names2df}{Default TRUE. If add miRNA and precusor names in the first two columns}

\item{message}{If print message}
}
\description{
Load miRDeep2 expression from a directory
}
\examples{

# 文件夹下有样本文件夹，样本表达为类似miRNAs_expressed_all_samples_06_09_2024.csv
dir.path="~/work/ProstaticCancer/analysis/321samples"
read_mirdeep2_result(dir.path)

}
