% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{reduce_join_list}
\alias{reduce_join_list}
\title{Join a table list}
\usage{
reduce_join_list(
  list.df,
  sfx = NULL,
  by = NULL,
  full = FALSE,
  inner = FALSE,
  left = FALSE,
  right = FALSE
)
}
\arguments{
\item{list.df}{List of data.frame}

\item{sfx}{If not specified, use list element names instead}

\item{by}{A character vector of variables to join by.}

\item{full}{If TRUE, use full_join}

\item{inner}{If TRUE, use inner_join}

\item{left}{If TRUE, use left_join}

\item{right}{If TRUE, use right_join}
}
\description{
Join a table list
}
