% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDMapping.R
\name{refSeq_Ensembl}
\alias{refSeq_Ensembl}
\title{Convert RefSeq ID to Ensembl ID by biomaRt}
\usage{
refSeq_Ensembl(nm.accessions)
}
\arguments{
\item{nm.accessions}{}
}
\value{

}
\description{
RefSeq identifiers differ in format according to the type of record the identifiers are for as shown below:
}
\details{
NG_XXXXX: RefSeq accessions for genomic region (nucleotide) records

NM_XXXXX: RefSeq accessions for mRNA records

NC_XXXXX: RefSeq accessions for chromosome records

NP_XXXXX: RefSeq accessions for protein records

XR_XXXXX: RefSeq accessions for model RNAs that are not associated with protein products

XM_XXXXX: RefSeq accessions for model mRNA records

XP_XXXXX: RefSeq accessions for model protein records

Where XXXXX is a sequence of integers.

NCBI https://www.ncbi.nlm.nih.gov/RefSeq/ allows users to query the RefSeq database using RefSeq identifiers.
}
