% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptsRelated.R
\name{remove.verylong.isoform.gff}
\alias{remove.verylong.isoform.gff}
\title{Remove isforms with extremly long interval (isoform.End - isoform.Start)}
\usage{
remove.verylong.isoform.gff(gff, interval.length = 1500000)
}
\arguments{
\item{gff}{gff file path}

\item{interval.length}{the maximum length between isoform start and isoform end}
}
\description{
Remove isforms with extremly long interval (isoform.End - isoform.Start)
}
\examples{
remove.verylong.isoform.gff("/data/home2/Zhongxu/work/3.filter.gtf")
}
