% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{removeBatchEffect}
\alias{removeBatchEffect}
\title{SVA combatc remove batch effect}
\usage{
removeBatchEffect(df = NULL, batch = NULL, group = NULL, consider.group = T)
}
\arguments{
\item{df}{row is gene, col is sample}

\item{consider.group}{Model matrix for outcome of interest and other covariates besides batch}
}
\description{
SVA combatc remove batch effect
}
