% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immune.R
\name{return_immune_genes}
\alias{return_immune_genes}
\title{Immune related gene}
\usage{
return_immune_genes()
}
\value{
A tibble object
}
\description{
Immune related gene
}
\examples{
loonR::return_immune_genes()
}
