% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{ridge_plot}
\alias{ridge_plot}
\title{Beautiful ridge plot}
\usage{
ridge_plot(
  term = NULL,
  value = NULL,
  logP = NULL,
  xlab = "Log2FoldChange",
  Legend = "-Log10(pvalue)",
  ylab = ""
)
}
\arguments{
\item{ylab}{}
}
\description{
Beautiful ridge plot
}
\examples{
term = sample(LETTERS[1:10], 100, TRUE)
value = rnorm(100,0,1)
}
