% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{riskCalibrationPlot.default}
\alias{riskCalibrationPlot.default}
\title{Calibration plot}
\usage{
\method{riskCalibrationPlot}{default}(
  group,
  pred,
  rms.method = FALSE,
  title = "Calibration plot",
  show.oberved.ci = FALSE,
  bins = 10,
  color = "npg",
  show.group = FALSE,
  ticks.unit = 0.25,
  full.range = TRUE
)
}
\arguments{
\item{group}{Must be a TRUE/FALSE factor}

\item{pred}{predicted probability}

\item{rms.method}{If TRUE, use rms::val.prob function instead}

\item{bins}{Number of bins. Default 20}

\item{color}{Default npg}

\item{show.group}{whether to plot histogram by group}

\item{ticks.unit}{0.25 seq(0, 1, by = 0.25)}

\item{full.range}{Default TRUE. loess smoothing between 0-1 or first bin to last bin}
}
\value{

}
\description{
Calibration plot
}
\examples{
data(BreastCancer)
BreastCancer = BreastCancer[,-c(1)]
BreastCancer = na.omit(BreastCancer)
m <- glm(Class ~ ., data = BreastCancer, family = binomial)
BreastCancer$pred <- predict(m, type = "response")
riskCalibrationPlot(factor(BreastCancer$Class=="malignant", levels=c(FALSE, TRUE)),
                   BreastCancer$pred)

data(LIRI)

d1 <- LIRI[,-c(1,5)]
m <- glm(status ~ ., data = d1, family = binomial(logit))
d1$pred <- predict(m, type = "response")
riskCalibrationPlot(factor(LIRI$status), d1$pred)
}
