% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{riskCalibrationPlot.lrm}
\alias{riskCalibrationPlot.lrm}
\title{Calibration plot by rms.calibrate}
\usage{
\method{riskCalibrationPlot}{lrm}(rms.model, cox = FALSE)
}
\arguments{
\item{rms.model}{A model built by rms package}

\item{cox}{If is a cox model}
}
\value{

}
\description{
Calibration plot by rms.calibrate
}
\examples{
#' data(LIRI)
m=loonR::build.logistic.model(LIRI[,c(3,4)],LIRI$status, rms = T)
m=m$model
riskCalibrationPlot(m)
}
