% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{riskCalibrationPlot}
\alias{riskCalibrationPlot}
\title{riskCalibrationPlot}
\usage{
riskCalibrationPlot(
  group,
  pred,
  rms.method = FALSE,
  title = "Calibration plot",
  show.oberved.ci = FALSE,
  bins = 10,
  color = "npg",
  show.group = FALSE,
  ticks.unit = 0.25,
  full.range = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
riskCalibrationPlot
}
