% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCA.R
\name{riskCalibrationPlotMultiple}
\alias{riskCalibrationPlotMultiple}
\title{Plot multiple calibration plot by riskRegression}
\usage{
riskCalibrationPlotMultiple(risk.list = NULL, label = NULL, palette = "aaas")
}
\arguments{
\item{risk.list}{Data.frame with column risk or a named list}

\item{label}{a vector}

\item{palette}{}
}
\description{
Plot multiple calibration plot by riskRegression
}
\examples{

data("LIRI")
risk.list = list(ANLN  = LIRI$ANLN,
                 CENPA = LIRI$CENPA)
label = LIRI$status

risk.list = LIRI[,c(3,4)]

}
