% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCA.R
\name{riskCalibrationPlotSingleComplex}
\alias{riskCalibrationPlotSingleComplex}
\title{Calibration plot}
\usage{
riskCalibrationPlotSingleComplex(risk, label, color = "#4682B4", bins = 10)
}
\arguments{
\item{risk}{Should be risk probability}

\item{bins}{Default 10}
}
\value{

}
\description{
Calibration plot
}
\examples{
data("LIRI")
risk = LIRI$ANLN
risk = loonR::logit2prob(scale(risk))
label = LIRI$status
loonR::riskCalibrationPlotSingleComplex(risk, label)

}
