% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{run_cellChat}
\alias{run_cellChat}
\title{Run cellchat}
\usage{
run_cellChat(seurat.obj, type = "triMean", min.cells = 100, cores = 20)
}
\arguments{
\item{min.cells}{minimum 100 for analysis}

\item{cores}{Default 20}
}
\value{
cellChat.Object
}
\description{
Run cellchat
}
