% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{run_RCTD}
\alias{run_RCTD}
\title{run_RCTD}
\usage{
run_RCTD(
  scRNA = NULL,
  spatial = NULL,
  rctd_mode = "doublet",
  max_multi_types = 10
)
}
\arguments{
\item{rctd_mode}{RCTD has three modes, each suited for different spatial technologies: "doublet": Assigns 1-2 cell types per pixel. Best for high-resolution technologies like Slide-seq and MERFISH. "multi": Like doublet mode but can assign more cell types (up to max_multi_types). Best for lower-resolution technologies like 100-micron Visium. "full": No restrictions on number of cell types.}
}
\description{
run_RCTD
}
