% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNF.R
\name{run_SNF}
\alias{run_SNF}
\title{Title Similarity network fusion}
\usage{
run_SNF(
  dataL = NULL,
  alpha = 0.5,
  K = 20,
  Iterations = 20,
  dist.method = "Euclidean Distance"
)
}
\arguments{
\item{dataL}{list( t(mRNA.snf.df), t(methylation.snf.df), t(cnv.snf.df) )}

\item{alpha}{Default 0.5. hyperparameter, usually (0.3~0.8)   Variance for local model}

\item{K}{Default 20. Number of neighbors, must be greater than 1. usually (10~30)  20}

\item{Iterations}{T.Default 20. Number of Iterations, usually (10~50)}

\item{dist.method}{"Euclidean Distance", Pearson, Spearman, mutualInfo}
}
\value{

}
\description{
Title Similarity network fusion
}
\examples{
run_SNF( list( t(mRNA.snf.df), t(methylation.snf.df), t(cnv.snf.df) ),  alpha = 0.5, K = 20, Iterations = 20    )
https://cran.r-project.org/web/packages/SNFtool/SNFtool.pdf
}
