% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{run_spotlight}
\alias{run_spotlight}
\title{run spotlight}
\usage{
run_spotlight(
  scRNA = NULL,
  spatial = NULL,
  downsample = 100,
  feature.auc = 0.75,
  top_n_feature = 50
)
}
\arguments{
\item{scRNA}{Seurat object}

\item{spatial}{Seurat object}

\item{downsample}{Default 100}

\item{feature.auc}{Default 0.75 when selecting cluster marker}

\item{top_n_feature}{Default 50}
}
\description{
run spotlight
}
