% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immune.R
\name{runEPIC}
\alias{runEPIC}
\title{run EPIC}
\usage{
runEPIC(expression.df)
}
\arguments{
\item{expression.df}{a matrix of the TPM (or RPKM) gene expression from the samples for which to estimate cell proportions}
}
\value{
Data frame with immune infiltration information
}
\description{
run EPIC
}
