% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immune.R
\name{runEstimate}
\alias{runEstimate}
\title{Run estiamte}
\usage{
runEstimate(expression.df, platform = c("affymetrix", "agilent", "illumina"))
}
\arguments{
\item{expression.df}{the row name in an input data must be gene symbols}

\item{platform}{c("affymetrix", "agilent", "illumina")}
}
\value{

}
\description{
Run estiamte
}
