% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Immune.R
\name{runImmunedeconv}
\alias{runImmunedeconv}
\title{Run Immunedeconv}
\usage{
runImmunedeconv(
  expression.df,
  method = c("quantiseq", "timer", "cibersort", "cibersort_abs", "mcp_counter", "xcell",
    "epic"),
  indication = ""
)
}
\arguments{
\item{expression.df}{gene × sample. TPM-normalized, not log-transformed}

\item{method}{c("quantiseq", "timer", "cibersort", "cibersort_abs", "mcp_counter", "xcell", "epic")}

\item{indication}{When use timer, should be one of immunedeconv::timer_available_cancers}
}
\description{
Run Immunedeconv
}
