% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{runInferCNV}
\alias{runInferCNV}
\title{Run inferCNV}
\usage{
runInferCNV(
  seurat.obj,
  ref_group_names = NULL,
  gene_pos = "~/ref/hg38/gencode.v44/human_genes_pos.txt",
  save_local_path = NULL
)
}
\arguments{
\item{ref_group_names}{Reference cell type}

\item{gene_pos}{gene order file}

\item{save_local_path}{Path to save rds}
}
\description{
Run inferCNV
}
