% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{runMetaFlux}
\alias{runMetaFlux}
\title{Run metaflux pipeline}
\usage{
runMetaFlux(
  seurat.obj = NULL,
  myident = NULL,
  n_bootstrap = 100,
  seed = 1,
  medium = NULL
)
}
\arguments{
\item{medium}{}
}
\description{
Run metaflux pipeline
}
