% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{runSlingshot}
\alias{runSlingshot}
\title{Run pesudo analysis}
\usage{
runSlingshot(seu.obj, start.clus = NULL, approx_points = 150)
}
\arguments{
\item{seu.obj}{Seurat object}

\item{start.clus}{NULL}

\item{approx_points}{150}
}
\value{
slingshot object with multiple function variable
}
\description{
Run pesudo analysis
}
\examples{
# Pls run runSlingshot first
# Next run f1, f2 function
# Then run runTradeSeq in the backend and save to continue
# Final run f4 and f5
}
