% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network.R
\name{runWGCNA}
\alias{runWGCNA}
\title{WGCNA, Weighted correlation network analysis}
\usage{
runWGCNA(
  datExpr,
  datTraits,
  traitColumnIndex = 1,
  power = NULL,
  minModuleSize = 30,
  MEDissThres = 0.25,
  nSelectHeatmap = 10000
)
}
\arguments{
\item{datExpr}{row is sample}

\item{datTraits}{0 1 to indicate group. Numeric to indicate trait}

\item{traitColumnIndex}{Trait we are interested in}

\item{power}{Default by sft$powerEstimate}

\item{minModuleSize}{We like large modules, so we set the minimum module size relatively high:}

\item{MEDissThres}{We choose a height cut of 0.25, corresponding to correlation of 0.75, to merge}

\item{nSelectHeatmap}{n (default 1000) genes to show in heatmap (a way to visualize network)}
}
\value{

}
\description{
WGCNA, Weighted correlation network analysis
}
