% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{scenicHeatmap}
\alias{scenicHeatmap}
\title{AUC heatmap for scenic}
\usage{
scenicHeatmap(auc.file = NULL, seu.obj = NULL, color = NULL)
}
\arguments{
\item{seu.obj}{}
}
\description{
AUC heatmap for scenic
}
\examples{
loonR::scenicHeatmap(auc.file.path, seu.obj)
}
