% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{SCT_Harmony_cluster}
\alias{SCT_Harmony_cluster}
\title{SCTranform HarmonyIntegration FindCluster}
\usage{
SCT_Harmony_cluster(
  seurat.obj,
  vars.to.regress = NULL,
  group.by.vars = "orig.ident",
  dims = 1:20,
  runSCT = T,
  runHarmony = T
)
}
\arguments{
\item{vars.to.regress}{colnames to remove covariate}

\item{group.by.vars}{colnames to remove batch effect}
}
\description{
SCTranform HarmonyIntegration FindCluster
}
