% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{scViolinPlot}
\alias{scViolinPlot}
\title{Similar to dotplot, use violin plot}
\usage{
scViolinPlot(object, groupBy, MarkerSelected, marker.group, color = NULL)
}
\arguments{
\item{object}{Seurat obj}

\item{groupBy}{Column names of identity}

\item{MarkerSelected}{marker gene names. vector}

\item{marker.group}{marker gene groups. Correspond with vector}

\item{color}{palatte}
}
\description{
Similar to dotplot, use violin plot
}
