% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{Seurat2Monocle3}
\alias{Seurat2Monocle3}
\title{convert seurat to monocle 3 obj}
\usage{
Seurat2Monocle3(
  seu.obj,
  partition.col.name = NULL,
  use_partition = F,
  auto.root = T,
  root.cluster.name = NULL,
  cluster.col.name = NULL
)
}
\arguments{
\item{auto.root}{if automatically identify root}

\item{root.cluster.name}{if auto.root is TRUE, set root cluster name}

\item{cluster.col.name}{if auto.root is TRUE, set cluster column name}
}
\description{
convert seurat to monocle 3 obj
}
