% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{splitCharacter}
\alias{splitCharacter}
\title{Splite vector and return the specific element}
\usage{
splitCharacter(vector, sep = NULL, index = 1)
}
\arguments{
\item{index}{}
}
\description{
Splite vector and return the specific element
}
