% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{splitDataByGroup}
\alias{splitDataByGroup}
\title{split data frame by group}
\usage{
splitDataByGroup(data.frame = NULL, group = NULL, seed = 666, fraction = 0.5)
}
\arguments{
\item{data.frame}{row is sample}

\item{seed}{666}

\item{fraction}{0.5}
}
\description{
split data frame by group
}
\examples{
data(LIRI)
res <- splitDataByGroup(data.frame = LIRI, group = LIRI$status)
}
