% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{splitGroupByCutoff}
\alias{splitGroupByCutoff}
\title{Split group by various methods. Useful}
\usage{
splitGroupByCutoff(
  group = NULL,
  values = NULL,
  fun = NULL,
  quantile.cutoff = NULL,
  sample.names = NULL,
  cut.point = NULL,
  cut.label = NULL,
  specific.group = NULL,
  group.prefix = NULL
)
}
\arguments{
\item{fun}{e.g. mean, median}

\item{quantile.cutoff}{0-1. percent value for quantile function}

\item{sample.names}{Default NA}

\item{cut.point}{Default NULL. Not include the minimum and maximum value}

\item{cut.label}{length should be the length(quantile.cutoff), or length(cut.point) plus 1}

\item{specific.group}{If specified, only split within these group}

\item{group.prefix}{""}
}
\value{

}
\description{
Split group by various methods. Useful
}
\examples{
set.seed(111)
value = runif(100, min=0, max=100)
set.seed(111)
group = sample(c(1,2,3),100, replace = TRUE)
table(group)

res <- loonR::splitGroupByCutoff(group, value, fun="mean", cut.label = c("L","H"), group.prefix = "G", specific.group = c(1,2))
table(res$New.Label)
}
