% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{splitSampleByGroup}
\alias{splitSampleByGroup}
\title{Split sample by group into traning and validation set}
\usage{
splitSampleByGroup(sample = NULL, group = NULL, seed = 666, fraction = 0.5)
}
\arguments{
\item{sample}{Vector}

\item{group}{Vector}

\item{seed}{Default 666}

\item{fraction}{Default 0.5}
}
\value{
list(Train=train.design, Validation=validation.design)
}
\description{
Split sample by group into traning and validation set
}
