% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suppa.R
\name{suppa.event.pie}
\alias{suppa.event.pie}
\title{Count each event type and draw pie chart}
\usage{
suppa.event.pie(
  data,
  color = "jco",
  colid = 2,
  alpha = 1,
  title = "",
  border = "white"
)
}
\arguments{
\item{data}{A data.frame or list object.}

\item{color}{ggsci color palette}

\item{colid}{If provide data.frame, column id need to be set}

\item{alpha}{Alpha value in plot}

\item{title}{Pie title}

\item{border}{Border color}
}
\value{

}
\description{
Count each event type and draw pie chart
}
\examples{
suppa.event.pie(ioe.events.df$Type, title = "# of events")
or suppa.event.pie(ioe.events.df, col = 2, title = "# of events")
}
