% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suppa.R
\name{suppa.get.final.table}
\alias{suppa.get.final.table}
\title{Combine dpis, psi and tpm file together
I modified the code from https://github.com/comprna/SUPPA/blob/master/scripts/Volcano_MA_plot.R}
\usage{
suppa.get.final.table(
  sample.names = "",
  psi = "",
  dpsi = "",
  event.tpm = "",
  pval.cutoff = 0.05,
  dpsi.cutoff = 0.3,
  tpm.cutoff = 0
)
}
\arguments{
\item{sample.names}{A vector specify the column name of each psi in psivec file.}

\item{psi}{Psivec file. Generate by Suppa2 diffSplice}

\item{dpsi}{Dpsi file. Generate by Suppa2 diffSplice file}

\item{event.tpm}{Event TPM file. Should be specified in Suppa analysis by --save_tpm_events}

\item{pval.cutoff}{P-value cutoff}

\item{dpsi.cutoff}{ΔPSI cutoff}

\item{tpm.cutoff}{Mean TPM cutoff}
}
\value{
A data.frame
}
\description{
Combine dpis, psi and tpm file together
I modified the code from https://github.com/comprna/SUPPA/blob/master/scripts/Volcano_MA_plot.R
}
\examples{

suppa.get.final.table(sample.names = smp.names,
         dpsi = "analysis/05suppa/diffSplice.events.dpsi.temp.0",
         psi  = "analysis/05suppa/diffSplice.events.psivec",
         event.tpm = "analysis/05suppa/diffSplice.events_avglogtpm.tab",
         dpsi.cutoff = dpsi.cutoff, pval.cutoff = pval.cutoff, tpm.cutoff = tpm.cutoff)

}
