% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suppa.R
\name{suppa.load.salmon.matrix}
\alias{suppa.load.salmon.matrix}
\title{Read Salmon output}
\usage{
suppa.load.salmon.matrix(dirPath)
}
\arguments{
\item{dirPath}{Simple set the directory which contains Salmon output folder}
}
\value{
A tximport oject
}
\description{
Read Salmon output
}
