% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppa.R
\name{SuppaEventPie}
\alias{SuppaEventPie}
\title{Count each event type and draw pie chart}
\usage{
SuppaEventPie(
  data,
  color = "jco",
  colid = 2,
  alpha = 1,
  title = "",
  border = "white"
)
}
\arguments{
\item{data}{A data.frame or list object.}

\item{color}{ggsci color palette}

\item{colid}{If provide data.frame, column id need to be set}

\item{alpha}{Alpha value in plot}

\item{title}{Pie title}

\item{border}{Border color}
}
\value{

}
\description{
Count each event type and draw pie chart
}
\examples{
SuppaEventPie(ioe.events.df$Type, title = "# of events")
or SuppaEventPie(ioe.events.df, col = 2, title = "# of events")
}
