% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppa.R
\name{SuppaLoadSalmonMatrix}
\alias{SuppaLoadSalmonMatrix}
\title{Read Salmon output}
\usage{
SuppaLoadSalmonMatrix(dirPath)
}
\arguments{
\item{dirPath}{Simple set the directory which contains Salmon output folder}
}
\value{
A tximport ojject
}
\description{
Read Salmon output
}
