% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{surv_ROC_DCA}
\alias{surv_ROC_DCA}
\title{Time depend ROC and AUC}
\usage{
surv_ROC_DCA(risk = NULL, event = NULL, time = NULL, timePoint = 12)
}
\arguments{
\item{timePoint}{}
}
\value{

}
\description{
Time depend ROC and AUC
}
\examples{
data(mayo, package = "survivalROC")
res = loonR::surv_ROC_DCA(mayo$mayoscore5, mayo$censor, mayo$time/30, timePoint = 24)
res$ROC
}
