% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{survivaly_analysis}
\alias{survivaly_analysis}
\title{Survival analysis}
\usage{
survivaly_analysis(
  Event = NULL,
  Time = NULL,
  Group = NULL,
  group.prefix = NA,
  ylab = "Survival probability",
  title = "",
  palette = "lancet",
  conf.int = FALSE,
  legend.position = "none",
  linetype = 1,
  calculate.pval = FALSE,
  only.consider.group = NULL,
  not.consider.group = NULL,
  risk.table = TRUE
)
}
\arguments{
\item{Event}{Required 0,1}

\item{Time}{Required}

\item{Group}{Required}

\item{group.prefix}{Default ""}

\item{title}{Default "Suvival analysis"}

\item{palette}{the color palette to be used.}

\item{conf.int}{logical value. If TRUE, plots confidence interval.}

\item{legend.position}{one of c("top", "bottom", "left", "right", "none"). Default is "none".}

\item{linetype}{line types. Allowed values includes i) "strata" for changing linetypes by strata (i.e. groups); ii) a numeric vector (e.g., c(1, 2)) or a character vector c("solid", "dashed").}

\item{calculate.pval}{If TRUE, just reture p value data.frame.}

\item{only.consider.group}{Groups to consider}

\item{not.consider.group}{Groups to exclude}

\item{risk.table}{Default TRUE. Show the strata Table}
}
\value{

}
\description{
Survival analysis
}
