% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{table2Figure}
\alias{table2Figure}
\title{Convert a text table to figure}
\usage{
table2Figure(
  data,
  rowname = TRUE,
  ttheme = "blank",
  top.black.line = 1:2,
  bottom.black.line = 1,
  subtitle = NULL,
  main.title = NULL,
  footnote = NULL,
  vline = NULL
)
}
\arguments{
\item{rowname}{Default TRUE}

\item{ttheme}{Default blank. character string the table style/theme. The available themes are illustrated in the ggtexttable-theme.pdf file. Allowed values include one of c("default", "blank", "classic", "minimal", "light", "lBlack", "lBlue", "lRed", "lGreen", "lViolet", "lCyan", "lOrange", "lBlackWhite", "lBlueWhite", "lRedWhite", "lGreenWhite", "lVioletWhite", "lCyanWhite", "lOrangeWhite", "mBlack", "mBlue", "mRed", "mGreen", "mViolet", "mCyan", "mOrange", "mBlackWhite", "mBlueWhite", "mRedWhite", "mGreenWhite", "mVioletWhite", "mCyanWhite", "mOrangeWhite" ). Note that, l = "light"; m = "medium"}

\item{top.black.line}{Default 1:2. Black line. From top}

\item{bottom.black.line}{Default 1. Black line From bottom}

\item{subtitle}{Subtitle if you want to show}

\item{main.title}{Main title if you want to show}

\item{footnote}{Italic footnote at botton if you want to show}

\item{vline}{}
}
\value{

}
\description{
Convert a text table to figure
}
\examples{
data(iris)
loonR::table2Figure( head(iris), main.title = "IRIS",  footnote = "Here is footnote")
}
