% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNF.R
\name{tanimoto_distance}
\alias{tanimoto_distance}
\title{Calculate tanimoto distance}
\usage{
tanimoto_distance(x, similarity = F)
}
\arguments{
\item{x}{matrix}

\item{similarity}{Default F. If true return similarity, else return distance}
}
\description{
Calculate tanimoto distance
}
