% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\name{tcgabiolinks.get.450k.methylation}
\alias{tcgabiolinks.get.450k.methylation}
\title{Download 450K methylation data}
\usage{
tcgabiolinks.get.450k.methylation(project)
}
\arguments{
\item{project}{e.g. TCGA-COAD. See getGDCprojects()$project_id}
}
\value{

}
\description{
Download 450K methylation data
}
