% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\name{tcgabiolinks.get.junction.coverage}
\alias{tcgabiolinks.get.junction.coverage}
\title{Download junction data by TCGAbiolinks. Pls note, V2 tag and Illumina HiSeq platform were selected}
\usage{
tcgabiolinks.get.junction.coverage(
  project,
  remove.Raw = FALSE,
  dir = "~/rspace/GDCdata",
  tag = "v2",
  platform = "Illumina HiSeq"
)
}
\arguments{
\item{project}{E.g. TCGA-LIHC}

\item{remove.Raw}{If to remove raw data}

\item{dir}{Raw data directory}

\item{tags}{V2}
}
\value{
list(clinical = clinical, expression = expression.df, group = group)
Group is a data.frame including label, short name and sample barcode
Clinical is a list
}
\description{
Download junction data by TCGAbiolinks. Pls note, V2 tag and Illumina HiSeq platform were selected
}
