% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{tcgabiolinks.get.RNA.expression.log2tpm}
\alias{tcgabiolinks.get.RNA.expression.log2tpm}
\title{Download RAN expression by TCGAbiolinks}
\usage{
tcgabiolinks.get.RNA.expression.log2tpm(project, remove.Raw = FALSE)
}
\arguments{
\item{project}{E.g. TCGA-LIHC}

\item{remove.Raw}{If to remove raw data}
}
\value{
list(clinical = clinical, expression = expression.df, group = group)
Expression was log2 transformed.
Group is a data.frame including label, short name and sample barcode
Clinical is a list
}
\description{
Download RAN expression by TCGAbiolinks
}
