% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCA.R
\name{time_serials_Cindex}
\alias{time_serials_Cindex}
\title{Draw C index across times}
\usage{
time_serials_Cindex(list.cox.model, df, palette = "aaas", main = "")
}
\arguments{
\item{df}{include time and status column names}

\item{palette}{aaas}

\item{main}{Default ""}
}
\description{
Draw C index across times
}
