% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{ttest_differential}
\alias{ttest_differential}
\title{Title}
\usage{
ttest_differential(
  df,
  group,
  cal.AUC = TRUE,
  exclude.zore = FALSE,
  alternative = "two.sided"
)
}
\arguments{
\item{alternative}{c("two.sided", "less", "greater")}
}
\value{

}
\description{
Title
}
\examples{
Calculate p value by t.test manually

}
