% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{univariate_cox}
\alias{univariate_cox}
\title{Title}
\usage{
univariate_cox(d.frame, status, time, scale = TRUE)
}
\arguments{
\item{d.frame}{Data.frame --- Row: sample, Column: gene expression}

\item{time}{OS, DFS, RFS et al.....}

\item{scale}{}
}
\value{

}
\description{
Title
}
