% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{univariate_cox_sthda}
\alias{univariate_cox_sthda}
\title{Title}
\usage{
univariate_cox_sthda(d.frame, status, time)
}
\arguments{
\item{d.frame}{Data.frame --- Row: sample, Column: gene expression}

\item{time}{OS, DFS, RFS et al.....}
}
\value{

}
\description{
Title
}
