% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{vector.group}
\alias{vector.group}
\title{Group a vector into list by group}
\usage{
vector.group(v, g)
}
\arguments{
\item{v}{vector}

\item{g}{group}
}
\description{
Group a vector into list by group
}
\examples{
loonR::vector.group( v = 1:10, g = rep(c(FALSE,TRUE),c(5,5)) )
}
