% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano Plot}
\usage{
volcano_plot(
  x,
  y,
  xlab = "Log2 Fold Change",
  ylab = "-log10(Adjusted P)",
  lg2fc = 1,
  p = 0.05,
  restrict.vector = NA,
  label = NA
)
}
\arguments{
\item{x}{Log2 Fold Change}

\item{y}{Adjusted P value}

\item{xlab}{Default "Log2"}

\item{ylab}{Default "-log10(Adjusted P)"}

\item{lg2fc}{Default cufoff 1}

\item{p}{Default cutoff 0.05}

\item{restrict.vector}{A TURE/FALSE factor. Only show TRUE point in the vector.}
}
\value{

}
\description{
Volcano Plot
}
