% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{volcano_plot_V2}
\alias{volcano_plot_V2}
\title{Volcano plot}
\usage{
volcano_plot_V2(
  logFC = NULL,
  adjusted.p = NULL,
  gene.names = NULL,
  p.cutoff = 0.05,
  logFC.cutoff = 1,
  show.top.n = 5,
  sig.genes = NULL,
  title = "Volcano Plot",
  subtitle = NULL,
  xlim = c(-3, 3)
)
}
\arguments{
\item{logFC}{Required}

\item{adjusted.p}{Required}

\item{gene.names}{Required}

\item{p.cutoff}{0.05}

\item{logFC.cutoff}{1}

\item{show.top.n}{5}

\item{sig.genes}{specific genes to show}

\item{title}{Default Volcano Plot}

\item{subtitle}{Default NULL}

\item{xlim}{}
}
\description{
Volcano plot
}
