% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptsRelated.R
\name{writeGff}
\alias{writeGff}
\title{Write to gff file}
\usage{
writeGff(x, gff.path)
}
\arguments{
\item{x}{Granges object.}

\item{gff.path}{Output file path.}
}
\value{

}
\description{
Write to gff file
}
