% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{AllEqual}
\alias{AllEqual}
\title{Check if all values in a vector are the same}
\usage{
AllEqual(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Check if all values in a vector are the same
}
\examples{

}
