% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{biomarker.discovery.pipeline}
\alias{biomarker.discovery.pipeline}
\title{Return heatmpa, cv-roc, average risk score}
\usage{
biomarker.discovery.pipeline(
  normal.sample,
  tumor.sample,
  expression.exp,
  candidate.miRNA = c(),
  AUC.cutoff = 0.8,
  FC.cutoff = 1.5,
  P.cutoff = 0.05,
  Exp.cutoff = -1,
  k = 2,
  n = 100,
  heatmap.group = "Cancer",
  donwregulated.biomarker = FALSE,
  scale = FALSE
)
}
\arguments{
\item{expression.exp}{No log2 transformation}

\item{candidate.miRNA}{Default c()}

\item{AUC.cutoff}{Default 0.8}

\item{FC.cutoff}{Default 1.5}

\item{P.cutoff}{Default 0.05}

\item{k}{Default 2}

\item{n}{Default 100}

\item{scale}{Please note. Default FALSE}
}
\value{
A list
}
\description{
Return heatmpa, cv-roc, average risk score
}
\examples{
biomarker.discovery.pipeline(samsung.healthy.exosome.samples, samsung.pdac.exosome.samples, all.sample.normlized.exp.df)
}
