% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{build.coxregression.model}
\alias{build.coxregression.model}
\title{Build cox regression model}
\usage{
build.coxregression.model(d.frame, status, time, seed = 666, scale = TRUE)
}
\arguments{
\item{d.frame}{Data.frame --- Row: sample, Column: gene expression}

\item{time}{OS, DFS, RFS et al.....}

\item{seed}{Default 666}

\item{scale}{}
}
\description{
Build cox regression model
}
\examples{
# psm (parametric survival model) uses a survival model based on functions and their parameters.
# cph (Cox Proportional Hazards Model and Extensions) is using the cox model (and the Anderson-Gill model) which is based on the hazard functions.

data(LIRI)
res = loonR::build.coxregression.model(LIRI[,3:5],LIRI$status, LIRI$time)
}
