% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{build.logistic.model}
\alias{build.logistic.model}
\title{Title}
\usage{
build.logistic.model(df, group, seed = 666, scale = TRUE)
}
\arguments{
\item{df}{Column is gene/miRNA, Row is sample}

\item{group}{Two levels. Second unique variable is defined as experiment group}
}
\value{

}
\description{
Title
}
