% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{build.logistic.model}
\alias{build.logistic.model}
\title{Build a logistic model}
\usage{
build.logistic.model(
  df,
  group,
  seed = 666,
  scale = TRUE,
  direction = "backward"
)
}
\arguments{
\item{df}{Column is gene/miRNA, Row is sample}

\item{group}{Two levels. Second unique variable is defined as experiment group}

\item{direction}{backward  c("both", "backward", "forward")}
}
\value{
A list. list(model=glm.fit,
StepwiseModel=elimination,
eliminationCandidates=stringr::str_remove_all(names(unclass(coef(elimination))\link{-c(1)}),'`')
)
}
\description{
Build a logistic model
}
