% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drug.R
\name{cellline_drug_gene_correlation}
\alias{cellline_drug_gene_correlation}
\title{Drug and gene correlation analysis}
\usage{
cellline_drug_gene_correlation(drug.sen.df = NULL, gene.expr = NULL)
}
\arguments{
\item{drug.sen.df}{A data.frame. Row is drug, col is cell line}

\item{gene.expr}{A vector with name. Name is cell line}
}
\value{

}
\description{
Drug and gene correlation analysis
}
