% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE.R
\name{ClusterProfiler.GSEA}
\alias{ClusterProfiler.GSEA}
\title{Perform GSEA analysis by cluster profiler}
\usage{
ClusterProfiler.GSEA(
  phenotype,
  geneNames,
  minGSSize = 10,
  qvalue = 0.05,
  GO = TRUE,
  KEGG = TRUE,
  MSigDb = TRUE,
  Hallmark = TRUE,
  exp.gene.type = "SYMBOL"
)
}
\arguments{
\item{phenotype}{log2FC}

\item{geneNames}{gene symbols. Must be symbol}

\item{minGSSize}{10}

\item{qvalue}{0.05}

\item{GO}{TRUE}

\item{KEGG}{TRUE}

\item{MSigDb}{TRUE}

\item{Hallmark}{TRUE}

\item{exp.gene.type}{RNA expression ID ENSEMBL. keytypes(org.Hs.eg.db)}
}
\description{
Perform GSEA analysis by cluster profiler
}
