% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{cross.validation}
\alias{cross.validation}
\title{k fold n times cross validation}
\usage{
cross.validation(
  df = NULL,
  label = NULL,
  k = 5,
  n = 100,
  scale = TRUE,
  type = "response"
)
}
\arguments{
\item{df}{data.frame. Row is sample, col is gene}

\item{label}{True label}

\item{k}{Folds}

\item{n}{Repeat times}

\item{scale}{TRUE}

\item{type}{"response" or "link"}
}
\description{
k fold n times cross validation
}
\examples{
cv.res <- loonR::cross.validation(miR.df[, cf.candidates],
            group == "Cancer",
            k = 10, n = 100)
}
