% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCA.R
\name{decisionCurveAnalysis}
\alias{decisionCurveAnalysis}
\title{Decision curve analysis. Similar with plot.ntbft}
\usage{
decisionCurveAnalysis(
  data.frame.list = NULL,
  label = NULL,
  rms = FALSE,
  validation.df = NULL,
  palette = "aaas"
)
}
\arguments{
\item{data.frame.list}{row is sample. List should inlcude names}

\item{rms}{Default FALSE}

\item{validation.df}{should include all variable in data.frame.list}

\item{palette}{}
}
\description{
Decision curve analysis. Similar with plot.ntbft
}
\examples{
data(LIRI)

d1 <- data.frame( LIRI[,c(3)])
d2 <- LIRI[,c(3,4)]
d3 <- LIRI[,c(3,4,5)]
data.frame.list = list(d1=d1,d2=d2,d3=d3)
decisionCurveAnalysis(data.frame.list, label=LIRI$status)

}
