% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{drawScatter}
\alias{drawScatter}
\title{Draw scatter plot}
\usage{
drawScatter(
  xvalue,
  yvalue,
  xlab = "X",
  ylab = "Y",
  group = NA,
  color = "jco",
  title = "",
  remove.legend = FALSE,
  margin = TRUE,
  xlim = NULL,
  ylim = NULL,
  show.sample.name = FALSE,
  label = NULL,
  cor.coef = F,
  cor.method = "pearson"
)
}
\arguments{
\item{xvalue}{Vector of values}

\item{yvalue}{The same length wtih x value}

\item{color}{Default jco}

\item{remove.legend}{Default FALSE}

\item{margin}{Default TRUE, show margin plot}

\item{show.sample.name}{Default FALSE}

\item{label}{Default NULL}

\item{cor.coef}{TRUE/FALSE to show coefficient}

\item{cor.method}{method for computing correlation coefficient. Allowed values are one of "pearson", "kendall", or "spearman"}
}
\value{
ggplot2 object
}
\description{
Draw scatter plot
}
\examples{
library(ggpubr)
data("mtcars")
loonR::drawScatter(mtcars$wt, mtcars$mpg, xlab = "wt", ylab = "mpg",  remove.legend = T, cor.coef = F)
}
