% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{drawScatter}
\alias{drawScatter}
\title{Draw scatter plot}
\usage{
drawScatter(
  xvalue,
  yvalue,
  xlab = "X",
  ylab = "Y",
  group = NA,
  color = "jco",
  title = "",
  margin = TRUE
)
}
\arguments{
\item{xvalue}{Vector of values}

\item{yvalue}{The same length wtih x value}

\item{color}{Default jco}

\item{margin}{Default TRUE, show margin plot}
}
\value{

}
\description{
Draw scatter plot
}
