% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA.R
\name{extractTCGASampleType}
\alias{extractTCGASampleType}
\title{Extract TCGA sample types}
\usage{
extractTCGASampleType(TCGA.ids)
}
\arguments{
\item{TCGA.ids}{}
}
\value{

}
\description{
Extract TCGA sample types
}
