% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{findSurvivalCutPoint}
\alias{findSurvivalCutPoint}
\title{Determine the Optimal Cutpoint for Continuous Variables}
\usage{
findSurvivalCutPoint(
  values = NULL,
  event = NULL,
  time = NULL,
  plot.Surv = FALSE
)
}
\arguments{
\item{time}{}
}
\description{
Determine the Optimal Cutpoint for Continuous Variables
}
\examples{

data(LIRI)
res = loonR::findSurvivalCutPoint(LIRI[,3],LIRI$status, LIRI$time)
res$pval
res$surv.plot
res$estimated.point
}
