% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorRelated.R
\name{get.ggsci.color}
\alias{get.ggsci.color}
\title{Get colors by ggsci}
\usage{
get.ggsci.color(palette = "nrc", alpha = 1, n = 10)
}
\arguments{
\item{palette}{default: nrc. Palette ref: https://nanx.me/ggsci/articles/ggsci.html}

\item{alpha}{1}

\item{n}{number of colors, default n = 10}
}
\value{

}
\description{
Get colors by ggsci
}
