% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{get_performance}
\alias{get_performance}
\title{Title Get model performace, sensitivity, sepcificity and others}
\usage{
get_performance(pred, labels, best.cutoff = NA)
}
\arguments{
\item{labels}{}
}
\value{

}
\description{
Title Get model performace, sensitivity, sepcificity and others
}
\examples{
get_performance(risk.probability, label)
}
