% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{getOneRoundCVRes}
\alias{getOneRoundCVRes}
\title{Title}
\usage{
getOneRoundCVRes(df, label, k, seed = 1)
}
\arguments{
\item{df}{data.frame. Col is sample, row is gene}

\item{k}{Folds}
}
\value{

}
\description{
Title
}
