% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{getOneRoundCVRes}
\alias{getOneRoundCVRes}
\title{Title}
\usage{
getOneRoundCVRes(df, label, k, seed = 1, times = 1)
}
\arguments{
\item{df}{data.frame. Row is sample, col is gene}

\item{label}{TRUE/FALSE vector}

\item{k}{Folds}

\item{times}{Default: 1. the number of pieces was use to predict. E.g. times = 5, will run 5 times, each time use a different piece to predict.}
}
\value{

}
\description{
Title
}
