% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptsRelated.R
\name{getTranscriptProteinSequence}
\alias{getTranscriptProteinSequence}
\title{Retrieve protein sequence by biomaRt}
\usage{
getTranscriptProteinSequence(IDs, type = "ensembl_transcript_id")
}
\arguments{
\item{type}{Default: ensembl_transcript_id, Available: ensembl_transcript_id, ensembl_transcript_id_version, refseq_mrna, Possible filters are given by the listFilters function. biomaRt::listFilters(mart)}
}
\value{

}
\description{
Retrieve protein sequence by biomaRt
}
