% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{heatmap.with.lgfold.riskpro}
\alias{heatmap.with.lgfold.riskpro}
\title{Plot heatmap with log 2 fold change and risk probability information}
\usage{
heatmap.with.lgfold.riskpro(
  heatmap.df,
  label,
  lgfold,
  risk.pro,
  scale = TRUE,
  group.name = "Cancer"
)
}
\arguments{
\item{heatmap.df}{Row: miRNA, Column: Sample}

\item{label}{A factor with labels TURE/FALSE}

\item{lgfold}{corresponded to row name sequence}

\item{group.name}{Default "Cancer"}

\item{risk.procorresponded}{to sample sequence}
}
\value{

}
\description{
Plot heatmap with log 2 fold change and risk probability information
}
\examples{
heatmap.with.lgfold.riskpro(data.tmp[candi,],label, logfd,  risk.pro)
}
