% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{hyperGeoTest}
\alias{hyperGeoTest}
\title{Perform hypergeometric test}
\usage{
hyperGeoTest(
  row.group,
  col.group,
  row.prefix = "",
  col.prefix = "",
  lower.tail = FALSE,
  title = "",
  log10.lowest = 5
)
}
\arguments{
\item{lower.tail}{Default FALSE}

\item{log10.lowest}{Default 3. Minimux or maximum log10 value. Useful when meet inf or draw heatmap}
}
\value{

}
\description{
Perform hypergeometric test
}
\examples{
g1 = sample(c("G1","G2","G3"),10, replace = T)
g2 = sample(c("1","2","3"),10, replace = T)
loonR::hyperGeoTest(g1, g2, col.prefix = "E")

}
