% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{lasso.select.feature}
\alias{lasso.select.feature}
\title{Feature selection by lasso}
\usage{
lasso.select.feature(
  data.matrix,
  label,
  folds = 5,
  seed = 666,
  family = "binomial",
  type.measure = "auc",
  s = NULL,
  scale = TRUE
)
}
\arguments{
\item{data.matrix}{Row is sample}

\item{family}{Default binomial}

\item{type.measure}{class, auc, deviance, mae. “deviance” uses actual deviance. “mae” uses mean absolute error. “class” gives misclassification error. “auc” (for two-class logistic regression ONLY) gives area under the ROC curve.}

\item{s}{Defalut is lambda.min. User can specify}

\item{scale}{Default TRUE}
}
\value{

}
\description{
Feature selection by lasso
}
