% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{limma.differential.cv.selection}
\alias{limma.differential.cv.selection}
\title{Perform multiple rounds differential analysis to select stable feature}
\usage{
limma.differential.cv.selection(
  log.df,
  label,
  folds = 5,
  seed = 666,
  n = 1000,
  cores = 50,
  AUC.cut.off = 0.8,
  FoldC.cut.off = 1
)
}
\arguments{
\item{log.df}{Row is gene, col is samples}

\item{label}{Normal first}

\item{folds}{5}

\item{seed}{666}

\item{n}{1000}

\item{cores}{50}

\item{AUC.cut.off}{0.8}

\item{FoldC.cut.off}{1}
}
\description{
Perform multiple rounds differential analysis to select stable feature
}
