% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{load.salmon.matrix}
\alias{load.salmon.matrix}
\title{Read Salmon output}
\usage{
load.salmon.matrix(
  dirPath,
  isoform = TRUE,
  countsFromAbundance = "no",
  dir.pattern = ""
)
}
\arguments{
\item{dirPath}{Simple set the directory which contains Salmon output folder}

\item{isoform}{specify data type. isoforma specific or not.}

\item{countsFromAbundance}{countsFromAbundance = c("no", "scaledTPM", "lengthScaledTPM", "dtuScaledTPM"),}
}
\value{
A tximport oject
}
\description{
Read Salmon output
}
\examples{
Directory tree: dir/sample_quant/quant.sf
}
