% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methylation.R
\name{loadMethylationArraryData}
\alias{loadMethylationArraryData}
\title{Load methylation arrary from directory}
\usage{
loadMethylationArraryData(
  dir,
  arraytype = "EPIC",
  combat = FALSE,
  batchname = c("Slide")
)
}
\arguments{
\item{arraytype}{EPIC or 450K}

\item{combat}{If run combat to remove batch effect}

\item{batchname}{Default c("Slide")}
}
\value{

}
\description{
Load methylation arrary from directory
}
\examples{
Reference: https://bioconductor.org/packages/release/bioc/vignettes/ChAMP/inst/doc/ChAMP.html
}
