% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{MA_plot}
\alias{MA_plot}
\title{M-A plot}
\usage{
MA_plot(
  M,
  A,
  p,
  m.cutoff = 0,
  a.cutoff = 0,
  p.cutoff = 0.05,
  mlab = "Log2 Fold Change",
  alab = "Average Expression",
  restrict.vector = NA,
  label = NA,
  col.pal = c("blue", "gray", "red")
)
}
\arguments{
\item{M}{Fold change or log ratio}

\item{A}{Average exprssion}

\item{p}{Adjusted P}

\item{m.cutoff}{Default 0}

\item{a.cutoff}{Default 0}

\item{p.cutoff}{Default 0.05}

\item{mlab}{Default 'Log2 Fold Change'}

\item{alab}{Default 'Average Expression'}

\item{restrict.vector}{A TURE/FALSE factor. Only show TRUE point in the vector.}

\item{label}{Point names which you want to show in plot. If you don't want to show, set NA}

\item{col.pal}{Default: c("blue", "gray", "red"). A vector with 3 elements. For significantly down, not significant, significantly up}
}
\description{
M-A plot
}
\examples{
MA_plot(tissue.exp.df.res$logFC, tissue.exp.df.res$AveExpr, tissue.exp.df.res$adj.P.Val)
}
