% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{meltDataFrameByGroup}
\alias{meltDataFrameByGroup}
\title{Melt data frame by group}
\usage{
meltDataFrameByGroup(d.frame, group, na.rm = TRUE, variable_name = "Gene")
}
\arguments{
\item{d.frame}{Row is sample, column is feature.}

\item{na.rm}{Default TRUE. Should NA values be removed from the data set?}
}
\value{

}
\description{
Melt data frame by group
}
\examples{
data(LIRI)
d.frame = LIRI[,3:6]
group = LIRI$status
loonR::meltDataFrameByGroup(d.frame, group)
}
