% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{multi_roc_with_ci}
\alias{multi_roc_with_ci}
\title{Plot multiple ROCs in one figure}
\usage{
multi_roc_with_ci(
  scores,
  labels,
  font = "Arial",
  palette = "jama",
  legend.pos = c(0.4, 0.2),
  title = NULL,
  panel = NULL,
  color = NULL,
  ci = TRUE
)
}
\arguments{
\item{scores}{A list or a data.frame. If list, labels shoule also be a list}

\item{color}{specify the color manually}
}
\value{

}
\description{
Plot multiple ROCs in one figure
}
\examples{
multi_roc_with_ci(rss, labels, font = "Arial", palette = "jama")
}
