% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{multivariate_or}
\alias{multivariate_or}
\title{Variate logistic analysis
Row: sample, Column: gene expression}
\usage{
multivariate_or(d.frame, label)
}
\arguments{
\item{d.frame}{Data.frame --- Row: sample, Column: gene expression}

\item{label}{True label}
}
\value{

}
\description{
Variate logistic analysis
Row: sample, Column: gene expression
}
