% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{plot_miRCorrelation}
\alias{plot_miRCorrelation}
\title{Plot miR's correlation}
\usage{
plot_miRCorrelation(df)
}
\arguments{
\item{df}{Row: miR expression, Column: Sample}
}
\value{

}
\description{
Plot miR's correlation
}
