% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Waterfall plot}
\usage{
plot_waterfall(
  risk.score,
  label,
  xlab = "Risk probability",
  palette = "jco",
  title = "",
  yticks.labl = c(0, 0.25, 0.5, 0.75, 1),
  sample = NA,
  rotate.x = 90
)
}
\arguments{
\item{xlab}{Risk probability}

\item{yticks.labl}{seq(0,1,by = 0.25) c(0,0.25,0.5,0.75,1)}

\item{sample}{If you want to show sample label, provide a vector}

\item{rotate.x}{Default 90. Rotate x axis text.}
}
\description{
Waterfall plot
}
\examples{
loonR::plot_waterfall(average.riskscore$Mean, average.riskscore$Label, xlab = "Risk probability")
}
