% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{plotBarWithErr}
\alias{plotBarWithErr}
\title{Plot bar with mean and standard error}
\usage{
plotBarWithErr(
  values,
  group,
  title = "",
  xlab = "X label",
  ylab = "Mean",
  color = "aaas",
  comparisons = "",
  method = "wilcox.test",
  label.y = NULL,
  rotate.x = 0
)
}
\arguments{
\item{values}{vector}

\item{group}{vector}

\item{comparisons}{list( c("N", "T") )}

\item{method}{wilcox.test or t.test}

\item{rotate.x}{Default 0. numeric value specifying the rotation angle. 90 for vertical x-axis text.}
}
\value{

}
\description{
Plot bar with mean and standard error
}
