% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{plotBarWithErr}
\alias{plotBarWithErr}
\title{Plot bar with mean and standard error}
\usage{
plotBarWithErr(
  values,
  group,
  title = "",
  xlab = "X label",
  ylab = "Mean",
  color = "aaas"
)
}
\arguments{
\item{values}{vector}

\item{group}{vector}

\item{color}{}
}
\value{

}
\description{
Plot bar with mean and standard error
}
