% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{plotClevelandDot}
\alias{plotClevelandDot}
\title{Draw Celventland dot plot}
\usage{
plotClevelandDot(
  name,
  value,
  group = NA,
  palette = "aaas",
  dot.size = 2,
  ylab = NULL,
  legend.title = "Group",
  title = "",
  cleveland = TRUE,
  lollipop = FALSE,
  value.name = "Value",
  sample.name = "Name"
)
}
\arguments{
\item{name}{Names}

\item{value}{values}

\item{group}{group}

\item{palette}{Default aaas, if group not set, use #00AFBB}

\item{dot.size}{Default 2}

\item{legend.title}{Default "Group"}

\item{title}{Default ""}

\item{value.name}{xlab, Default "Value"}

\item{sample.name}{ylab, Default "Name"}
}
\value{

}
\description{
Draw Celventland dot plot
}
