% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{plotJitterBoxplot}
\alias{plotJitterBoxplot}
\title{Boxplot with jitter}
\usage{
plotJitterBoxplot(
  values,
  group,
  title = "",
  xlab = "Group",
  ylab = "Value",
  color = "aaas",
  comparisons = "",
  method = "wilcox.test",
  label.y = NULL,
  add = "jitter",
  alternative = "both"
)
}
\arguments{
\item{values}{vector}

\item{group}{vector}

\item{comparisons}{list( c("N", "T") )}

\item{method}{wilcox.test or t.test}

\item{label.y}{}
}
\value{

}
\description{
Boxplot with jitter
}
