% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{plotPCA}
\alias{plotPCA}
\title{PCA plot
Please note: scale will be performed autmaticly}
\usage{
plotPCA(
  df,
  group,
  palette = "npg",
  ellipse = FALSE,
  legend.title = "Class",
  main.title = "",
  alpha = 1,
  return.percentage = FALSE,
  pre.filter = 0.01,
  label = NULL
)
}
\arguments{
\item{df}{Row: sample, Column: gene expression}

\item{palette}{Default npg. The color palette to be used for coloring or filling by groups.
Allowed values include "grey" for grey color palettes; brewer palettes e.g. "RdBu", "Blues", ...;
or custom color palette e.g. c("blue", "red");
and scientific journal palettes from ggsci R package, e.g.: "npg", "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".}

\item{ellipse}{Whether to add elipse.}

\item{legend.title}{legend.title}

\item{main.title}{main title}

\item{return.percentage}{If TRUE, reture PCA percentage instead of PCA lot}

\item{lable}{Must be a vector or NULL}
}
\value{

}
\description{
PCA plot
Please note: scale will be performed autmaticly
}
\examples{
plotPCA(df, group, "aaas")
}
