% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{plotPie}
\alias{plotPie}
\title{Count each event type and draw pie chart}
\usage{
plotPie(
  data,
  color = "jco",
  colid = 2,
  alpha = 1,
  title = "",
  border = "white",
  label = FALSE,
  show.total.inTitle = FALSE,
  DonutChart = FALSE
)
}
\arguments{
\item{data}{A data.frame or list object.}

\item{color}{ggsci color palette}

\item{colid}{If provide data.frame, column id need to be set}

\item{alpha}{Alpha value in plot}

\item{title}{Pie title}

\item{border}{Border color}

\item{label}{Whether to show labels}

\item{show.total.inTitle}{If to show the total number in title}

\item{DonutChart}{If plot based on donut chart. More beautiful.}
}
\value{

}
\description{
Count each event type and draw pie chart
}
\examples{
plotPie(ioe.events.df$Type, title = "# of events")
or plotPie(ioe.events.df, col = 2, title = "# of events")

plotPie( rep(c("A", "B", "C"), c(10, 60, 30)) )


}
