% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{plotSilhouette}
\alias{plotSilhouette}
\title{Title}
\usage{
plotSilhouette(
  df,
  group,
  color = "aaas",
  class = "Class",
  label = FALSE,
  alpha = 0.8
)
}
\arguments{
\item{df}{row is feature, column is sample}

\item{label}{Default FALSE}
}
\value{

}
\description{
Title
}
