% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA.R
\name{read_mirdeep2_result}
\alias{read_mirdeep2_result}
\title{Load miRDeep2 expression matrix}
\usage{
read_mirdeep2_result(file.path, add.names2df = TRUE)
}
\arguments{
\item{file.path}{mirdeep2 expression. Should be merged before}

\item{add.names2df}{Default TRUE. If add miRNA and precusor names in the first two columns}
}
\description{
Load miRDeep2 expression matrix
}
