% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{riskCalibrationPlot}
\alias{riskCalibrationPlot}
\title{Calibration plot}
\usage{
riskCalibrationPlot(
  group,
  pred,
  smooth = TRUE,
  title = "Calibration plot",
  show.oberved.ci = FALSE
)
}
\arguments{
\item{group}{Must be a TRUE/FALSE factor}

\item{pred}{predicted probability}

\item{smooth}{Default TRUE}

\item{show.oberved.ci}{}
}
\value{

}
\description{
Calibration plot
}
\examples{
data(BreastCancer)
BreastCancer = BreastCancer[,-c(1)]
BreastCancer = na.omit(BreastCancer)
m <- glm(Class ~ ., data = BreastCancer, family = binomial)
BreastCancer$pred <- predict(m, type = "response")
riskCalibrationPlot(as.factor(BreastCancer$Class=="malignant", levels=c(FALSE, TRUE)),
                   BreastCancer$pred)

}
