% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{riskCalibrationPlot}
\alias{riskCalibrationPlot}
\title{Calibration plot}
\usage{
riskCalibrationPlot(
  group,
  pred,
  smooth = TRUE,
  title = "Calibration plot",
  show.oberved.ci = FALSE,
  bins = 10,
  color = "npg",
  show.group = FALSE,
  ticks.unit = 0.25,
  full.range = TRUE
)
}
\arguments{
\item{group}{Must be a TRUE/FALSE factor}

\item{pred}{predicted probability}

\item{smooth}{Default TRUE}

\item{bins}{Number of bins. Default 20}

\item{color}{Default npg}

\item{show.group}{whether to plot histogram by group}

\item{ticks.unit}{0.25 seq(0, 1, by = 0.25)}

\item{full.range}{Default TRUE. loess smoothing between 0-1 or first bin to last bin}
}
\value{

}
\description{
Calibration plot
}
\examples{
riskCalibrationPlot(factor(BreastCancer$Class=="malignant", levels=c(FALSE, TRUE)),
                   BreastCancer$pred)
}
