% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biomarker.R
\name{roc_with_ci}
\alias{roc_with_ci}
\title{plot ROC with CI. Stratified bootstrap 2000 times}
\usage{
roc_with_ci(
  label,
  rs,
  font = "Arial",
  palette = "jama",
  legend.pos = c(0.4, 0.2),
  title = NULL,
  fontsize = 16,
  panel = NULL
)
}
\arguments{
\item{label}{True label}

\item{rs}{Predicted score}

\item{font}{Font}

\item{panel}{panel的第一列为factor，event，class等}
}
\value{
ggplot object
}
\description{
plot ROC with CI. Stratified bootstrap 2000 times
}
\examples{
roc_with_ci(labels, rs,
font = "Arial",
palette = "jama_classic",
title = "HGD vs Healthy",
panel = data.frame(Evebt=labels, data)
)
}
