% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{RTN.analysis}
\alias{RTN.analysis}
\title{RTN Reconstruction of Transcriptional regulatory Networks analysis}
\usage{
RTN.analysis(
  TFs = NULL,
  expData = NULL,
  group = NULL,
  cores = 1,
  nPermutations = 1000,
  phenotype = NULL,
  hits = NULL
)
}
\arguments{
\item{TFs}{transcription factors. vector}

\item{expData}{row is gene}

\item{cores}{Default 1}

\item{nPermutations}{Default 1000}

\item{phenotype}{A named vector}

\item{hits}{}
}
\value{

}
\description{
RTN Reconstruction of Transcriptional regulatory Networks analysis
}
