% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{scaleDF}
\alias{scaleDF}
\title{Scale a data.frame by column or row}
\usage{
scaleDF(df, byRow = FALSE, byColumn = TRUE, center = TRUE, scale = TRUE)
}
\arguments{
\item{byRow}{Default FALSE}

\item{byColumn}{Default TRUE, by column}

\item{center}{Default TRUE. Mean = 0}

\item{scale}{Default TRUE. 0-1 scale}
}
\value{

}
\description{
Scale a data.frame by column or row
}
