% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNF.R
\name{SNF_Similairity_Hist}
\alias{SNF_Similairity_Hist}
\title{Similairity count}
\usage{
SNF_Similairity_Hist(affinityL = NULL, evidence.type = NULL, group = NULL)
}
\arguments{
\item{affinityL}{List including different affinity matrix}

\item{evidence.type}{Corresponded to affinity list}

\item{group}{SNF subtyping result}
}
\value{

}
\description{
Similairity count
}
\examples{
loonR::SNF_Similairity_Hist(tmp$AffinityL,evidence.type = c("RNA","Methylation", "CNV"), group = snf.group)
}
