% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{survivaly_analysis}
\alias{survivaly_analysis}
\title{Survival analysis}
\usage{
survivaly_analysis(
  Event = NULL,
  Time = NULL,
  Group = NULL,
  group.prefix = NA,
  ylab = "Survival probability",
  title = "",
  palette = "lancet",
  conf.int = FALSE,
  legend.position = "none",
  linetype = 1,
  calculate.pval = FALSE,
  remove.na = FALSE,
  only.consider.group = NULL,
  not.consider.group = NULL,
  risk.table = TRUE,
  best.point = F,
  cut.quantile = NULL,
  cut.label = NULL,
  surv.median.line = "none",
  pval = TRUE,
  returnSurv.fit = FALSE,
  maximum.days = NULL
)
}
\arguments{
\item{Event}{Required 0,1}

\item{Time}{Required}

\item{Group}{Required}

\item{group.prefix}{Default ""}

\item{title}{Default "Suvival analysis"}

\item{palette}{the color palette to be used.}

\item{conf.int}{logical value. If TRUE, plots confidence interval.}

\item{legend.position}{one of c("top", "bottom", "left", "right", "none"). Default is "none".}

\item{linetype}{line types. Allowed values includes i) "strata" for changing linetypes by strata (i.e. groups); ii) a numeric vector (e.g., c(1, 2)) or a character vector c("solid", "dashed").}

\item{calculate.pval}{If TRUE, just reture p value data.frame.}

\item{remove.na}{If remove NA samples}

\item{only.consider.group}{Groups to consider}

\item{not.consider.group}{Groups to exclude}

\item{risk.table}{Default TRUE. Show the strata Table}

\item{best.point}{If best cut point wanted, pls input values by Group variable}

\item{cut.quantile}{Love this paramter, handsome}

\item{surv.median.line}{character vector for drawing a horizontal/vertical line at median survival. Allowed values include one of c("none", "hv", "h", "v"). v: vertical, h:horizontal.}

\item{pval}{logical value, a numeric or a string or show "HR" or "PHR" or "HRCI" or "PHRCI"}

\item{returnSurv.fit}{If TRUE, reture the fit object. Easy to check median survival time}

\item{maximum.days}{The maximum days or months}
}
\value{

}
\description{
Survival analysis
}
\examples{
data(LIRI)
loonR::survivaly_analysis(LIRI$status, LIRI$time, LIRI$ANLN > mean(LIRI$ANLN), legend.position="right", risk.table = F )
loonR::survivaly_analysis(LIRI$status, LIRI$time, LIRI$ANLN, best.point = T)
}
