% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{ttest_differential}
\alias{ttest_differential}
\title{Calculate p value by t.test manually}
\usage{
ttest_differential(
  df,
  group,
  cal.AUC = TRUE,
  exclude.zore = FALSE,
  exclude.na = TRUE,
  alternative = "two.sided",
  cores = 40,
  paired = FALSE
)
}
\arguments{
\item{df}{Row is gene, column is sample}

\item{alternative}{c("two.sided", "less", "greater")}

\item{paired}{Default FALSE}
}
\description{
Calculate p value by t.test manually
}
\examples{
Parametric Tests of Group Differences
}
