% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{unique_gene_expression}
\alias{unique_gene_expression}
\title{Title Remove redundant gene expression data, and select the maximum one}
\usage{
unique_gene_expression(expression.df)
}
\arguments{
\item{expression.df}{Please note the first column must be gene names}
}
\value{
A clean expression data.frame
}
\description{
Title Remove redundant gene expression data, and select the maximum one
}
\examples{
loonR::unique_gene_expression(normlized.exp.df)
}
